### 2021-08-04 #3560 @whedon
[PRE REVIEW]: CleanX: A Python library for data cleaning of large sets of radiology images
**Submitting author:** @drcandacemakedamoore  (<a href="http://orcid.org/0000-0003-1672-7565">Candace Makeda Moore</a>)
**Repository:** <a href="https://github.com/drcandacemakedamoore/cleanX" target ="_blank">https://github.com/drcandacemakedamoore/cleanX</a>
**Version:** v0.1.9

Thanks for submitting your paper to JOSS @drcandacemakedamoore . **Currently, there isn't an JOSS editor assigned** to your paper.
