### 2021-08-04 9e9a77c
Minor changes to paper.md to aid pdf compilation
Minor changes to paper.md to aid pdf compilation
