+++ paper.md
@@ -5,30 +5,22 @@ tags:
   - radiology
   - imaging
   - medical imaging
-
 authors:
-  - name: Candace Makeda Moore, MD
+  - name: Candace Makeda Moore
     orcid: 0000-0003-1672-7565
     affiliation: 1,3
-
-  - name: Andrew Murphy, MMIS BMedImagingSc RT(R)
+  - name: Andrew Murphy
     orcid: 0000-0002-7710-6095
     affiliation: 2    
-
   - name: Oleg Sivokon
     affiliation: 3  
-     
-
 affiliations:
  - name: MaxCor lab, Sapir College, Israel
    index: 1
-
  - name: Department of Medical Imaging, Princess Alexandra Hospital, Brisbane, QLD, Australia
    index: 2   
-
  - name: CMHM/cmhm.info, Israel
    index: 3   
-
 date: 18 April 2021
 bibliography: paper.bib
 ---
