### 2021-08-04 #3560 @Kevin-Mattheus-Moerman
@drcandacemakedamoore your paper failed to compile. I think this is because of the use of comma's in the author names.

I think this pull-request should fix it:
https://github.com/drcandacemakedamoore/cleanX/pull/30

Let me know if you'd prefer to have the `, MD` etc. part back in, then we'll trouble shoot this further to try to include it.
