### 2021-08-04 d51a1f5
Affiliation shouldn't be a list, maybe a string then?
