+++ paper.md
@@ -10,7 +10,7 @@ tags:
 authors:
   - name: Candace Makeda Moore
     orcid: 0000-0003-1672-7565
-    affiliation: [1, 3]
+    affiliation: "1,3"
   - name: Andrew Murphy, MMIS BMedImagingSc RT(R)
     orcid: 0000-0002-7710-6095
     affiliation: 2
