+++ binder/requirements.txt
@@ -1,0 +1,7 @@
+
+cleanX==0.1.9
+numpy==1.18.5
+matplotlib==3.4.2
+pandas==1.3.0
+cv2==4.5.0
+PIL==8.3.0
