+++ cleanX/dataset_processing/dataframes.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 """
-clean X : Library for cleaning radiological data used in machine learning
+Library for cleaning radiological data used in machine learning
 applications.
+
 Module dataframes: a module for processing of datasets related to images.
 This module can be implemented by functions,
 or can be implemented with classes.
@@ -12,7 +13,7 @@ import html
 import logging
 import re
 
-from abc import ABC
+from abc import ABC, abstractmethod
 from collections.abc import Iterable, Sequence
 from pathlib import Path
 
@@ -22,12 +23,28 @@ import textwrap as tw
 
 
 class GuesserError(TypeError):
+    """
+    This error is raised when the loading code cannot figure
+    out what kind of source it is dealing with.
+
+    """
     pass
 
 
 class ColumnsSource(ABC):
+    """
+    Formal superclass for all sources that should be used to
+    produce :class:`~pandas.DataFrame`.
+    """
 
+    @abstractmethod
     def to_dataframe(self):
+        """
+        Descendants of this class must implement this method.
+
+        :return: Dataframe produced from the source represented by this class.
+        :rtype: :class:`~pandas.DataFrame`.
+        """
         raise NotImplementedError()
 
 
@@ -35,43 +52,117 @@ class CSVSource(ColumnsSource):
     """Class that helps turn csv into a dataframe"""
 
     def __init__(self, csv, **pd_args):
+        """
+        Initializes this class with the path to the csv file from which
+        to create the dataframe and the pass-through arguments to
+        :func:`pandas.read_csv()`.
+
+        :param csv: The path to CSV file or an open file handle.  Must be
+                    suitable for :func:`pandas.read_csv()`.
+        :param \\**pd_args: The pass-through arguments to
+                            :func:`pandas.read_csv()`.
+        """
         self.csv = csv
         self.pd_args = pd_args or {}
 
     def to_dataframe(self):
+        """
+        Necessary implementation of the abstractmethod.
+
+        :return: Dataframe produced from the source represented by this class.
+        :rtype: :class:`~pandas.DataFrame`.
+        """
         return pd.read_csv(self.csv, **self.pd_args)
 
 
 class JSONSource(ColumnsSource):
     """Class that helps turn json into a dataframe for later exploration"""
+
     def __init__(self, json, **pd_args):
+        """
+        Initializes this class with the path to the json file from which
+        to create the dataframe and the pass-through arguments to
+        :func:`pandas.read_json()`.
+
+        :param json: The path to JSON file or an open file handle.  Must be
+                    suitable for :func:`pandas.read_json()`.
+        :param \\**pd_args: The pass-through arguments to
+                            :func:`pandas.read_json()`.
+        """
         self.json = json
         self.pd_args = pd_args or {}
 
     def to_dataframe(self):
+        """
+        Necessary implementation of the abstractmethod.
+
+        :return: Dataframe produced from the source represented by this class.
+        :rtype: :class:`~pandas.DataFrame`.
+        """
         return pd.read_json(self.json, **self.pd_args)
 
 
 class DFSource(ColumnsSource):
-    """Class initializes dataframes for later exploration and reporting"""
+    """
+    This class is a no-op source.  Use this when you already have a dataframe
+    ready.
+    """
+
     def __init__(self, df):
+        """
+        Initializes this class with the existing dataframe.
+
+        :param df: Existing dataframe that should be used as the source.
+        :type df: :class:`~pandas.DataFrame`.
+        """
         self.df = df
 
     def to_dataframe(self):
+        """
+        Necessary implementation of the abstractmethod.
+
+        :return: Dataframe produced from the source represented by this class.
[…]
