+++ binder/requirements.txt
@@ -4,4 +4,4 @@ numpy==1.18.5
 matplotlib==3.4.2
 pandas==1.3.0
 opencv-python-headless
-PIL
+Pillow
