### 2021-08-05 #3560 @Kevin-Mattheus-Moerman
Note for AEiC on call: If no alternative handling editor becomes available within 2 weeks I can take this one.
