### 2021-08-05 4eee521
Multiple formatting changes in documentation
