+++ cleanX/image_work/image_functions.py
@@ -49,11 +49,11 @@ def crop_np(image_array):
     Crops black edges of an image array
 
     :param image_array: Image array.
-    :type image_array: array
+    :type image_array: :class:`~numpy.ndarray`
 
 
     :return: NumPy array with the image data with the black margins cropped.
-    :rtype: numpy.ndarray
+    :rtype: :class:`~numpy.ndarray`
     """
     nonzero = np.nonzero(image_array)
     y_nonzero = nonzero[0]
@@ -109,10 +109,10 @@ def subtle_sharpie_enhance(image):
     by human eye.
 
     :param image: String for image name
-    :type image: string
+    :type image: str
 
     :return: new_image_array, a nearly imperceptibly sharpened image for humans
-    :rtype: numpy.ndarray
+    :rtype: :class:`~numpy.ndarray`
     """
     ksize = (2, 2)
     image_body = cv2.imread(image)
@@ -129,10 +129,10 @@ def harsh_sharpie_enhance(image):
     but should be used with care to make augmented data.
 
     :param image: String for image name
-    :type image: string
+    :type image: str
 
     :return: new_image_array, a sharpened image for humans
-    :rtype: numpy.ndarray
+    :rtype: :class:`~numpy.ndarray`
     """
     ksize = (20, 20)
     image_body = cv2.imread(image)
@@ -147,10 +147,10 @@ def salting(img):
     not been tested for similarity to older machines, which also add noise.
 
     :param img_name: String for image name
-    :type img_name: string
+    :type img_name: str
 
     :return: new_image_array, with noise
-    :rtype: nd.array
+    :rtype: :class:`~numpy.ndarray`
     """
     kernel = (5, 5)
     img = cv2.imread(img)
@@ -199,7 +199,7 @@ def blur_out_edges(image):
     :type image: Image (JPEG)
 
     :return: blurred_edge_image an array of an image blurred around the edges
-    :rtype: numpy.ndarray
+    :rtype: :class:`~numpy.ndarray`
     """
     example = cv2.imread(image, cv2.IMREAD_GRAYSCALE)
     msk = np.zeros(example.shape)
@@ -257,7 +257,7 @@ def show_major_lines_on_image(pic_name):
     lines more likely to be non-biological.
 
     :param pic_name: String of image full name e.g. "C:/folder/image.jpg"
-    :type pic_name: string
+    :type pic_name: str
 
     :return: shows image but technically returns a matplotlib plotted image
     :rtype: matplotlib.image.AxesImage
@@ -292,13 +292,13 @@ def find_big_lines(directory, line_length):
     are all contiguous
 
     :param directory: Directory with set_of_images (should include final '/').
-    :type directory: directory
+    :type directory: Suitable for :func:`os.path.join()`
     :param line_length: Minimal length of lines for the function to count.
-    :type line_length: integer
+    :type line_length: int
 
     :return: :code:`DataFrame` with column for line count at or above
              :code:`line_length`
-    :rtype: :pd:`DataFrame`
+    :rtype: :class:`~pandas.DataFrame`
     """
     suspects1 = glob.glob(os.path.join(directory, '*.[Jj][Pp][Gg]'))
     suspects2 = glob.glob(os.path.join(directory, '*.[Jj][Pp][Ee][Gg]'))
@@ -336,9 +336,10 @@ def separate_image_averager(set_of_images, s=5):
     To run on a list to make a prototype tiny X-ray that is an averages image
 
     :param set_of_images: Set_of_images
-    :type set_of_images:
+    :type set_of_images: Collection of elements suitable for
+                         :func:`os.path.join()`
     :param s: length of sides in the image made
-    :type s: integer
+    :type s: int
 
     :return: image
     :rtype: image
[…]
