+++ README.md
@@ -100,6 +100,6 @@ We are adding new functions and classes all the time. Many unit tests are availa
 
         Normalize images in terms of pixel values (multiple methods)
 
-All important functions are documented in the online documentation for programmers. You can also check out our video by clicking the linked picture below:
+All important functions are documented in the online documentation for programmers. You can also check out one of our videos by clicking the linked picture below:
 
 [![Video](https://raw.githubusercontent.com/drcandacemakedamoore/cleanX/main/test/cleanXpic.png)](https://youtu.be/jaX5tXmiWrQ)
