+++ source/index.rst
@@ -5,12 +5,16 @@ cleanX documentation
 
 Here is some general information about the cleanX package. We are working on
 documentation for programmers and medical professionals with some coding
-experience. 
-The package helps with data cleaning for two-dimensional medical images. The
-idea is that given DICOMs or files of images, you can use the functions as APIs
-to build the basis for machine learning or certain other types of algorithms.
+experience. For now documentation for professionals with less coding experience
+can be found on our wiki. The documentation on this website is programmers
+oriented.  
+The cleanX package helps with data cleaning for two-dimensional medical images.
+The idea is that given DICOMs or files of images, you can use the functions as 
+APIs to build the basis for machine learning or certain other types of
+algorithms.
 You can use the package code to clean out data that would lower the quality of
-your algorithm, and use the code to add augmented images.
+your algorithm if the "uncleaned" data was included, and use also the code to
+add augmented images.
 
 
 The API Documentation / Guide
