### 2021-08-07 2584f4a
new sizing functions linted, tested
