### 2021-08-07 #3560 @drcandacemakedamoore
Thanks, we are constantly adding things. Will the editors look only at the current released version? And if so, should we release a new version in about two weeks when you may review, and let you know?
