+++ cleanX/image_work/image_functions.py
@@ -1739,13 +1739,13 @@ def show_close_images(folder, compression_level, ref_mse):
                 plt.axis("off")
                 # show the images
                 plt.show()
-                print("Similar files: ", image_files[main_img], " and ", 
-                    image_files[compared_img],)
+                print("Similar files: ", image_files[main_img], " and ",
+                        image_files[compared_img],)
                 duplicates_A.append(image_files[main_img])
                 duplicates_B.append(image_files[compared_img])
                 err_list.append(spec_err)
                 dupers = {'twinA?': duplicates_A, 'twinB?': duplicates_B,
-                        'mse': err_list}
+                          'mse': err_list}
                 near_dupers = pd.DataFrame(dupers)
                 srow_B += nrows
                 erow_B += nrows
@@ -1759,10 +1759,10 @@ def show_close_images(folder, compression_level, ref_mse):
         compared_img = main_img + 1
 
     print("\n***\n Output: ", str(len(duplicates_A)),
-        " potential duplicate image pairs in ", str(len(image_files)),
-        " total images.\n",
-        "At compression level",
-        compression,
-        "and mse",
-        ref_mse,)
+            " potential duplicate image pairs in ", str(len(image_files)),
+            " total images.\n",
+            "At compression level",
+            compression,
+            "and mse",
+            ref_mse,)
     return near_dupers
