### 2021-08-07 caab80d
[fix] infinite loop in image comparison test
