### 2021-08-07 #3560 @cMadan
@drcandacemakedamoore, it sounds like you have some updates on the go. Ideally the JOSS reviewers will review a stable version of the code, suggest some improvements, and then a version with those changes will be the archived version that the JOSS publication is associated with. From your earlier question, I wonder if it would it be better to wait for before we start the review?
