### 2021-08-07 fc57d08
[fix] Wrong kind of checkout for documentation
