### 2021-08-08 678abcf
more fixing notebooks
