### 2021-08-08 #3560 @drcandacemakedamoore
I will freeze updating everything on the Main branch to make this easier. The only important changes were to dramatically improve the documentation (we found some problems that Sphinx had made us, and fixed them). I would like to move forward with the review as this is re-submission after substantial work, and we anticipate that work will continue for years as dependencies change.
