### 2021-08-08 7a91266
Prevent matplitlib interactive ui from appearing during test
