### 2021-08-08 abee13b
Remove inherited Exception members from documentation
