### 2021-08-09 3a58149
[fix] Minor fixes in documentation + adding IPython refs
