### 2021-08-09 95e8105
More documentation for DICOM adapters
