### 2021-08-09 c583428
More documentation for DICOM adapters
