### 2021-08-12 aa49d6b
minor changes to readme
