+++ README.md
@@ -38,7 +38,7 @@ other authors + contributors: Oleg Sivokon, Andrew Murphy
 
 Online documentation at https://drcandacemakedamoore.github.io/cleanX/
 
-We encourage you to build up-to-date documentation by command.
+You can also build up-to-date documentation by command.
 
 Documentation can be generated by command:
 
@@ -51,7 +51,9 @@ The documentation will be generated in `./build/sphinx/html` directory. Document
 automatically as new functions are added.
 
 Special additional documentation for medical professionals with limited programming
-ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation)
+ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
+To get a high level overview of some of the functionality of the program you
+can look at the Jupyter notebooks inside workflow_demo. 
 
 # Installation
 - setting up a virtual environment is desirable, but not absolutely necessary
