### 2021-08-18 #3632 @whedon
[REVIEW]: CleanX: A Python library for data cleaning of large sets of radiology images
**Submitting author:** @drcandacemakedamoore  (<a href="http://orcid.org/0000-0003-1672-7565">Candace Makeda Moore</a>)
**Repository:** https://github.com/drcandacemakedamoore/cleanX
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.14

- [x] **Contribution and authorship:** Has the submitting author (@drcandacemakedamoore ) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Contribution and authorship:** Has the submitting author (@drcandacemakedamoore ) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
