### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
Joss issues
This is a list of some improvements/suggestions or issues that may need clarifications.

- [x] Is this file needed GNU GENERAL PUBLIC LICENSE.txt?
- [x] Include Conda badges https://anaconda.org/doctormakeda/cleanx/badges
- [x] Make sure that the test badges link to the test builds. Currently, they link to the image of the badge. [![Sanity](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-commit.yml/badge.svg)](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-commit.yml)
- [x] Create a `paper` folder for the paper files and include a copy of the LICENSE file.
- [x] Include some examples on how to get started in the readme file. The same applies to the documentation. I would expect at least some sort of getting started guide.

- [x] Since version v0.1.9 was released, I would expect the current changes to have v0.2.0.dev as the version for these changes in development. Later to be released as v0.2.0. But if you desire to have the current pattern, thats fine.
- [x] Move all document files to a docs folder. I think `readthedocs` could also enable the docs have two versions, the stable and the latest.
- [x] In the `Jupyter` we have paths like `'D:/projects/cleanX'` It would be nice to start by getting the current project's directory and then use relative paths with join.
dicomfile_directory1 = 'D:/projects/cleanX/test/dicom_example_folder'
example = pd.read_csv("D:/projects/cleanX/workflow_demo/martians_2051.csv")
# To
working_dir = "Path to project home"
example_path = os.path.normpath(os.path.join(working_dir, "workflow_demo/martians_2051.csv"))
example = pd.read_csv(example_path)
```
It would be nice to normalize the paths. This will help Windows users who have a hard time with / and \ characters
