### 2021-08-21 #3632 @henrykironde
I have requested some suggestions as part of the review https://github.com/drcandacemakedamoore/cleanX/issues/36
https://github.com/drcandacemakedamoore/cleanX/issues/35
