### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
I'll see what I can do about the rest. Quick comment on documentation first: there probably won't be readthedocs integration in the near future due to how they require us to submit our code. There are multiple problems there, to name a few: we need to install shared libraries necessary for Tesseract to work, but that's not possible, because we have no control over the environment in which the documentation is built.

There is some way to mock dependencies that won't be installed, unfortunately, that would require modifying the `install_requires` part of `setup.py` in a way that doesn't list the actually required packages... bottom line: we tried, and readthedocs is just too hard to integrate with / not worth the effort. I just need to remove the shreds of this failed attempt.

Also, I'm not quite sure about what do you mean when you write about moving documentation to docs folder. API documentation is all generated... there isn't any special folder for it as it's published as GitHub pages, separately from the rest of the project. Technically, when it's generated, it will end up under `./build/sphinx/html/...` but I don't think this is relevant to anyone.
