### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
I totally understand the readmedocs problem in case where we are including shared libraries. For the part that of the docs folder, It would be nice to have all the doc resources in the docs folder. i.e  `Makefile`, `make.bat`, `source`
