### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
I also looked into publishing multiple versions of documentation... And yes, there's something we could do, but some other things require a lot of work. The problem here is that two Sphinx "extensions" that aspire to do this are, basically, broken. The first one was unmaintained since Python 3.5 and doesn't work with any older versions, the second one doesn't integrate with `setuptools`, which is how we build the documentation. Not only that, it still requires a lot of manual labor to add versions (so, basically, if you go down that path, it's probably easier to just do it yourself by editing Sphinx HTML templates).
