### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
> all the doc resources in the docs folder. i.e Makefile, make.bat, source

Oh, yeah... those. Well, Makefile and make.bat are irrelevant. I simply forgot to delete them. I can rename source to docs though.
