### 2021-08-21 drcandacemakedamoore/cleanX#36 @wvxvw
We don't actually test `cleanX` with Python from python.org on systems other than Linux. We only have Anaconda tests for them.

Similarly, on MS Windows, you'd have to find a way to install Tesseract shared library.

Neither me nor @drcandacemakedamoore have a Mac to test how it works. Anaconda was relatively simple to support, as it uses the same process for installing shared libraries on Linux and everywhere else. But I would need at least a VM running flavors of OSX and Windows to start to understand how to deal with these dependencies if not through Anaconda.

----

Bottom line, for now, if you want to run it on anything other than Linux, you either need to use Anaconda, or you are on your own installing Tesseract shared libraries.

I'd try this:

``` sh
brew install tesseract
```

And hope for the best. I don't actually know what it will do and whether it will work. On Linux, `tesseract` is a number of different packages, mostly because different languages ares split in their own packages. But, maybe for install to succeed the above will be enough, if it works.
