+++ README.md
@@ -13,6 +13,7 @@ JPEG files can be extracted from [DICOM](https://www.dicomstandard.org/) files o
 ### The latest official release:
 
 <a href="https://pypi.org/project/cleanX/"><img alt="PyPI" src="https://img.shields.io/pypi/v/cleanX"></a>
+[![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/version.svg)](https://anaconda.org/doctormakeda/cleanx)
 
 
 primary author: Candace Makeda H. Moore
@@ -31,7 +32,7 @@ other authors + contributors: Oleg Sivokon, Andrew Murphy
 - ability to create virtual environments (recommended, not absolutely necessary)
 - tesseract-ocr, matplotlib, pandas, pillow and opencv
 - optional recommendation of simpleITK or pydicom for DICOM/dcm to jpg conversion
-- anaconda is now supported, but not technically necessary
+- anaconda is now supported, but not technically necessary [![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/platforms.svg)](https://anaconda.org/doctormakeda/cleanx)
 
 
 ## Documentation
