+++ README.md
@@ -5,9 +5,10 @@
 # cleanX
 
 CleanX <a href="https://doi.org/10.5281/zenodo.4725904"><img src="https://zenodo.org/badge/DOI/10.5281/zenodo.4725904.svg" alt="(DOI)"></a> <a href="https://github.com/drcandacemakedamoore/cleanX/blob/master/LICENSE"><img alt="License: GPL-3" src="https://img.shields.io/github/license/drcandacemakedamoore/cleanX"></a>
-is an open source python library
+is an open source [![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/license.svg)](https://anaconda.org/doctormakeda/cleanx) python library
 for exploring, cleaning and augmenting large datasets of X-rays, or certain other types of radiological images.
 JPEG files can be extracted from [DICOM](https://www.dicomstandard.org/) files or used directly.
+[![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/platforms.svg)](https://anaconda.org/doctormakeda/cleanx)
 
 
 ### The latest official release:
@@ -32,7 +33,7 @@ other authors + contributors: Oleg Sivokon, Andrew Murphy
 - ability to create virtual environments (recommended, not absolutely necessary)
 - tesseract-ocr, matplotlib, pandas, pillow and opencv
 - optional recommendation of simpleITK or pydicom for DICOM/dcm to jpg conversion
-- anaconda is now supported, but not technically necessary [![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/platforms.svg)](https://anaconda.org/doctormakeda/cleanx)
+- anaconda is now supported, but not technically necessary 
 
 
 ## Documentation
