### 2021-08-21 drcandacemakedamoore/cleanX#36 @drcandacemakedamoore
I just noticed Anaconda seems to be on ou3 1.8 version. I will look into it and try to fix it to 1.9 today.
