### 2021-08-21 e50a0a8
take away anaconda version for now
