### 2021-08-21 c97982d
[fix] Accidentally removed the code to upload packages to Anaconda
