### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
@henrykironde

> Make sure that the test badges link to the test builds. Currently, they link to the image of the badge.

I just checked with documentation: https://docs.github.com/en/actions/managing-workflow-runs/adding-a-workflow-status-badge this is how it is supposed to work. I.e. the URL for the badge is the way the documentation tells us to do, specifically:

> You reference the workflow by the name of your workflow file.
>
>
> ``` markdown
> ![example workflow](https://github.com/<OWNER>/<REPOSITORY>/actions/workflows/<WORKFLOW_FILE>/badge.svg)
> ```

I don't believe there's any way to point to the latest GitHub Actions run, even the concrete URL is only available *after* the job starts, so I cannot add it to the README in the same commit that starts the CI job.
