### 2021-08-21 05f5447
[fix] Remove unnecessary documentation helpers, rename source dir
