+++ .readthedocs.yaml
@@ -1,15 +1,0 @@
-# .readthedocs.yaml
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-
-# Required
-version: 2
-
-# Build documentation in the docs/ directory with Sphinx
-sphinx:
-   configuration: source/conf.py
-python:
-  version: 3.7
-  install:
-    - method: setuptools
-      path: .
