### 2021-08-21 be644e4
[fix] Add setup.py command to upload to Anaconda and use it in CI
