+++ .github/workflows/on-tag.yml
@@ -62,8 +62,10 @@ jobs:
           python setup.py genconda -t ${{ matrix.python-version }}
       - run: conda build -c conda-forge --output-folder ./dist ./conda-pkg/
       - run: >
-          ANACONDA_API_TOKEN=${{ secrets.ANACONDA_TOKEN }}
-          anaconda upload --force --label main ./dist/*/*.tar.bz2
+          conda run -n cleanx-build
+          python setup.py anaconda_upload
+          --token ${{ secrets.ANACONDA_TOKEN }}
+          --package ./dist/*/*.tar.bz2
       - uses: actions/upload-artifact@v2
         with:
           name: conda-build-${{ matrix.os }}-${{ matrix.python-version }}
