### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
If you open your workflows you will be able to create the status badge the right way.
This will enable contributors and users to easily check out the builds.
