### 2021-08-21 drcandacemakedamoore/cleanX#36 @henrykironde
I would also add the system requirements to the readme file to let people know the easy way to install on the various supported OS
