### 2021-08-21 4d94c47
Move paper and related files into its own directory
