+++ .github/workflows/on-commit.yml
@@ -137,9 +137,9 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          paper-path: paper.md
+          paper-path: paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
           name: paper
-          path: paper.pdf
+          path: paper/paper.pdf
