### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
@henrykironde Yes, I tried it. It just links to the Actions tab of the project... it's not really useful either. Why does it make a difference? That won't help anyone to find their builds, unfortunately...
