### 2021-08-21 e63efbf
Removed duplicate license
