### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
> Since version v0.1.9 was released, I would expect the current changes to have v0.2.0.dev as the version for these changes in development. Later to be released as v0.2.0. But if you desire to have the current pattern, thats fine.

All the changes are cosmetic so far. They don't touch the code. I.e. there's no API changes. We use semantic versioning with the semantics:

1. Most significant component means backwards-incompatible changes to API.
2. Second most significant component means backwards-compatible changes to API.
3. Least significant component means no changes to API.

We don't believe in using anything other than digits in versions. This creates a lot of problems for dependency constraint solvers s.a. those in `pip` or `conda`. So we aren't going to have versions for `dev`, `alpha`, `rc` etc. That's intended.
