### 2021-08-21 d412635
Add some tables to the README
