### 2021-08-21 7780ec5
[fix] Try different table formatting
