### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
> Yes, I tried it. It just links to the Actions tab of the project... it's not really useful either. Why does it make a difference? That won't help anyone to find their builds, unfortunately...

The goal is to make sure the submission is as good as the dependencies that are in the project. Take an example of https://github.com/matplotlib/matplotlib. They provide the links to the builds in the right markdown way.
When it comes to dependencies, developers will usually check on the builds of the packages they are utilizing. In this way you can have a better idea when the build failed and why and how that failure transitions to your package.
