### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
> All the changes are cosmetic so far. They don't touch the code. I.e. there's no API changes. We use semantic versioning with the semantics:

There could be a slight disagreement on the way you version the package. In case you want to use this style of versioning, I recommend creating a new branch. Call this branch development or any name. Once a release is made, no commits should be made to the main.
All commits should be to the dev branch, and when it is time to release you merge the dev to master.
**Why this is a recommended case:**
My software downloads scientific packages and analyses/detect the changes in code by version. I always build from the source because I don't like already build binaries or installers. As I detect for changes and malware, your package has a release version and a hash value, after you change the code, the hash of the code changes but the version is still the same. Right there other people using your source may have problems.
