### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
For the paper folder the license should not have the postfix `_copy`.
