+++ README.md
@@ -29,11 +29,19 @@ other authors + contributors: Oleg Sivokon, Andrew Murphy
 
 ## Requirements
 
-- a [python](https://www.python.org/downloads/) installation (3.7, 3.8 or 3.9)
-- ability to create virtual environments (recommended, not absolutely necessary)
-- tesseract-ocr, matplotlib, pandas, pillow and opencv
-- optional recommendation of simpleITK or pydicom for DICOM/dcm to jpg conversion
-- anaconda is now supported, but not technically necessary
+- a [python](https://www.python.org/downloads/) installation (3.7, 3.8
+  or 3.9)
+- ability to create virtual environments (recommended, not absolutely
+  necessary)
+- [`tesserocr`](https://github.com/sirfz/tesserocr),
+  [`matplotlib`](https://matplotlib.org/),
+  [`pandas`](https://pandas.pydata.org/),
+  [`pillow`](https://python-pillow.org/) and
+  [`opencv`](https://opencv.org/)
+- optional recommendation of [`SimpleITK`](https://simpleitk.org/) or
+  [`pydicom`](https://github.com/pydicom/pydicom) for DICOM/dcm to JPG
+  conversion
+- Anaconda is now supported, but not technically necessary
 
 
 ### Supported Platforms
@@ -84,61 +92,223 @@ python setup.py apidoc
 python setup.py build_sphinx
 ```
 
-The documentation will be generated in `./build/sphinx/html` directory. Documentation is generated
-automatically as new functions are added.
+The documentation will be generated in `./build/sphinx/html`
+directory. Documentation is generated automatically as new functions
+are added.
 
-Special additional documentation for medical professionals with limited programming
-ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
-To get a high level overview of some of the functionality of the program you
-can look at the Jupyter notebooks inside workflow_demo. 
+Special additional documentation for medical professionals with
+limited programming ability is available on the wiki
+(https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
+
+To get a high level overview of some of the functionality of the
+program you can look at the Jupyter notebooks inside workflow_demo.
+
 
 # Installation
-- setting up a virtual environment is desirable, but not absolutely necessary
+- setting up a virtual environment is desirable, but not absolutely
+  necessary
 
 - activate  the environment
-### Anaconda Installation
+
+
+## Anaconda Installation
 
 - use command for conda as below
 
-        conda install -c doctormakeda -c conda-forge cleanx       
+``` sh
+conda install -c doctormakeda -c conda-forge cleanx
+```
 
 You need to specify both channels because there are some cleanX
 dependencies that exist in both Anaconda main channel and in
 conda-forge
 
-### pip installation
+
+## pip installation
+
 - use pip as below
 
-        pip install cleanX
-    
-    
+``` sh
+pip install cleanX
+```
 
-## About using this library
-If you use the library, please credit me and my collaborators.  You are only free to use this library according to license. We hope that if you use the library you will open source your entire code base, and send us modifications.  You can get in touch with me by starting a discussion (https://github.com/drcandacemakedamoore/cleanX/discussions/37) if you have a legitimate reason to use my library without open-sourcing your code base, or following other conditions, and I can make you specifically a different license.
+# Getting Started
 
-We are adding new functions and classes all the time. Many unit tests are available in the test folder. Test coverage is currently partial. Some newly added functions allow for rapid automated data augmentation (in ways that are realistic for radiological data). Some other classes and functions are for cleaning datasets including ones that: 
[…]
