### 2021-08-21 3e2bebe
[fix] Added the fix for missing old Tesseract
