+++ cleanX/image_work/image_functions.py
@@ -4,6 +4,8 @@ Library for cleaning radiological data used in machine learning
 applications
 """
 
+import subprocess
+
 # imported libraries
 import cv2
 import numpy as np
@@ -12,6 +14,29 @@ import pandas as pd
 
 from PIL import Image
 from PIL import Image, ImageOps
+
+try:
+    def __fix_tesserocr_locale():
+        output = subprocess.check_output(
+            ['ldconfig', '-v'],
+            stderr=subprocess.DEVNULL,
+        )
+        for line in output.decode().split('\n'):
+            if line.lstrip().startswith('libtesseract'):
+                alias, soname = line.strip().split(' -> ')
+                _, _, maj, min, patch = soname.split('.')
+                maj, min, patch = int(maj), int(min), int(patch)
+                if (maj, min, patch) < (4, 0, 1):
+                    import locale
+                    locale.setlocale(locale.LC_ALL, 'C')
+                    logging.warning(
+                        'Setting locale to C, otherwise tesseract segfaults',
+                    )
+    __fix_tesserocr_locale()
+    del __fix_tesserocr_locale
+except FileNotFoundError:
+    logging.warning('Don\'t know how to find Tesseract library version')
+
 from tesserocr import PyTessBaseAPI
 
 import glob
