### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
> Why this is a recommended case:
> My software downloads scientific packages and analyses/detect the changes in code by version. I always build from the source because I don't like already build binaries or installers. As I detect for changes and malware, your package has a release version and a hash value, after you change the code, the hash of the code changes but the version is still the same. Right there other people using your source may have problems.

People building from source chose their own fate. If they don't want to rely on the distributable package built by the package maintainers, it's their own choice. Which I can understand, but can accommodate on the basis of available free time and ease of use for other people involved with the project.

This package is mostly written by medical professional, who aspire to be software developers on the side. I'm not a medical professional, and, normally, I follow the practices you recommend. However, what you recommend is uncomfortable and confusing to people who aren't professional software developers.

There's another aspect to it: the CI-related code necessary to release versions may have bugs. It's unfortunate, but this is how these free services work. I cannot test CI code in advance and make sure that it will work on the first try. This means that releases will sometimes fail. There is no way to roll out a release other than from the main branch. So, the changes will have to be made to the main branch even if nothing in the released code changes.

Yet another aspect of it is the documentation. In order to see that documentation is generated properly one has to do it from the main branch. While it's possible to see most of the changes from side branches, ultimately, the proof that it works is by generating it from the main branch. So, anyone who works on documentation is bound to make changes to the main branch. Again, this is an unfortunate aspect of how GitHub Actions work, and how the free hosting of documentation works.

So, in an ideal world, where I'm self-hosting my repository and organise CI and documentation publishing in whatever way I want, I could have main branch dedicated to releases only. This is not possible if using GitHub though outside of trivial cases.
