### 2021-08-21 drcandacemakedamoore/cleanX#35 @wvxvw
> The goal is to make sure the submission is as good as the dependencies that are in the project. Take an example of https://github.com/matplotlib/matplotlib. They provide the links to the builds in the right markdown way.

1. Our links are in exactly the same format as theirs.
2. They are equally useless... I mean, they just point to the Actions tab.
