### 2021-08-21 abeda6e
[fix] subprocess throws different errors if it cannot find an executable
