### 2021-08-21 drcandacemakedamoore/cleanX#35 @henrykironde
Well I will disagree with some of your opinions, If I have sometime I could help. Documentation should be tested locally first. Then push to whichever branch, that is why we have stable docs and latest docs.

> However, what you recommend is uncomfortable and confusing to people who aren't professional software developers.

That is why we are here to help as much as possible so that you publish a state of art project.

>  I cannot test CI code in advance and make sure that it will work on the first try.

We can change the workflow.
 1. Do not commit directly to the main. This is a very good practice. make branches and create pull requests. The CI should pass to allow you to merge the changes to master. In that way we are sure to a good percentage that we shall have bug free code.
2.
> There is no way to roll out a release other than from the main branch.

Releases are always from the main branch. Basing on the way you version your software, if you only commit to a dev branch, at the time of release you merge the dev to main and create a release.

3. >  the proof that it works is by generating it from the main branch
We can have tests for the dev branch too.

Let me know when you get time, I could probably spare an hour to help out with some of the enhancement so that we can be on the same page.
