### 2021-08-22 drcandacemakedamoore/cleanX#36 @wvxvw
@henrykironde We've done it already. See starting here: https://github.com/drcandacemakedamoore/cleanX/blame/main/README.md#L47
