### 2021-08-22 drcandacemakedamoore/cleanX#35 @wvxvw
> Documentation should be tested locally first. Then push to whichever branch, that is why we have stable docs and latest docs.

In an ideal world: yes. On GitHub: no. Here's the reason: the version of the library is determined by the latest tag on the current branch. If you generate documentation from a branch other than main, you need to make sure the latest tag on that branch is the same as the one on main. This may not always be possible.

A significant portion of work needed to publish documentation is done in GitHub actions. They are impossible to test locally. Because it's a hosted proprietary service. You can imitate some of it, but you cannot run it exactly as it does on GitHub. Even to develop an imitation of these GitHub actions will be a huge waste of time, that is completely unjustified if compared to simplicity of just running them on GitHub and watching the results.

Just to illustrate my point better: the images built for CI in GitHub are some rather weird mix-and-mash of Linux distros. This, for example, creates unexpected problems when our code tries to work around Tesseract's bugs related to Unicode locale support: who could've imagine that `ldconfig` wouldn't be installed on Linux? And yet, it isn't. I don't have direct access to GitHub CI images, not to mention that they keep changing every few weeks, and I have no control over what's going to be inside those images next month or, perhaps, next week. If I don't test in actual GitHub Actions environment, I will not discover the problems.

Trust me, I'm used to running CI with dozens of thousands of jobs a day, supporting projects much bigger and more complex than this one. What I do is based on many years of experience of doing this. I can recognize the zeal of wanting to do things "the right way". But, the truth of it is that CI needs to serve the people who work on the project, and make *their* life more convenient. If "the right way" is inconvenient, it is the wrong way. No matter how much you want to believe the opposite.

> Do not commit directly to the main. This is a very good practice.

You can change workflow in your projects however you like... I already explained why workflow in this project is the way it is and will remain the way it is. Thank you for your suggestion, but I'm not interested in changing it.
