### 2021-08-22 drcandacemakedamoore/cleanX#36 @drcandacemakedamoore
Just a note on my earlier comment about bringing everything up to 1.9: in some hours today everything will go under the new versions 1.10 or perhaps 1.11, not 1.9 for technical reasons. We are trying to limit releases, but this may be unavoidable. Right now everything but Windows is up to 1.10. Whatever the case we should have everything the new version out in less than 24 hours now. Thank you for your patience reviewer.
