### 2021-08-22 3cf82a4
Split CLI code into submodules
