### 2021-08-22 92d7381
Add rip_out_jpgs to DICOM readers
