### 2021-08-22 2b3bfc5
Add test for image extraction, closes #40
