+++ test/test_cli.py
@@ -5,6 +5,7 @@ import os
 import subprocess
 
 from tempfile import TemporaryDirectory
+from glob import glob
 
 import pandas as pd
 import pytest
@@ -18,7 +19,7 @@ from harness import skip_if_missing
 
 
 @skip_if_missing('no pydicom available', 'pydicom', 'click')
-def test_cli_pydicom():
+def test_cli_pydicom_report():
     dicomfile_directory1 = os.path.join(
         os.path.dirname(__file__),
         'dicom_example_folder',
@@ -27,7 +28,7 @@ def test_cli_pydicom():
         result = subprocess.call(
             [
                 sys.executable, '-m', 'cleanX',
-                'dicom', 'extract-images',
+                'dicom', 'report',
                 '-i', 'dir', dicomfile_directory1,
                 '-o', td,
             ]
@@ -39,13 +40,47 @@ def test_cli_pydicom():
 
 
 @skip_if_missing('no pydicom available', 'pydicom', 'click')
+def test_cli_pydicom_extract():
+    dicomfile_directory1 = os.path.join(
+        os.path.dirname(__file__),
+        'dicom_example_folder',
+    )
+    with TemporaryDirectory() as td:
+        result = subprocess.call(
+            [
+                sys.executable, '-m', 'cleanX',
+                'dicom', 'extract',
+                '-i', 'dir', dicomfile_directory1,
+                '-o', td,
+            ]
+        )
+        assert not result
+
+        jpegs = sorted(glob(os.path.join(td, '*.jpg')))
+        dicoms = sorted(os.listdir(dicomfile_directory1))
+        assert len(jpegs) == len(dicoms)
+
+        mismatches = []
+        for jpg, dcm in zip(jpegs, dicoms):
+            jpg = os.path.basename(jpg)
+            jpg = os.path.splitext(jpg)[0]
+
+            dcm = os.path.basename(dcm)
+            dcm = os.path.splitext(dcm)[0]
+
+            if jpg != dcm:
+                mismatches.append((jpg, dcm))
+        assert not mismatches
+
+
+@skip_if_missing('no pydicom available', 'pydicom', 'click')
 def test_cli_datasets():
     resources = os.path.dirname(__file__)
     with TemporaryDirectory() as td:
         result = subprocess.check_output(
             [
                 sys.executable, '-m', 'cleanX',
-                'dataset', 'generate-report',
+                'dataset', 'report',
                 '-r', os.path.join(resources, 'test_sample_df.csv'),
                 '-t', os.path.join(resources, 'train_sample_df.csv'),
                 # These two sets don't appear to have common columns
