+++ cleanX/dicom_processing/simpleitk_adapter.py
@@ -282,7 +282,7 @@ class SimpleITKDicomReader:
         :type source: :class:`~cleanX.dicom_processing.Source`
         :param destination: The name of the directory where JPG files
                             should be stored.
-        :type destination: Compatible with :fn:`os.path.join`
+        :type destination: Compatible with :func:`os.path.join`
         """
         reader = sitk.ImageFileReader()
         m_reader = MetadataHelper(reader)
