+++ cleanX/cli/__init__.py
@@ -9,5 +9,3 @@ from .main import main
 from .dicom import dicom, extract, report
 from .images import images, run_pipeline, restore_pipeline
 from .dataset import dataset, report
-
-
