### 2021-08-22 drcandacemakedamoore/cleanX#35 @wvxvw
>  Include some examples on how to get started in the readme file. The same applies to the documentation. I would expect at least some sort of getting started guide.

There are now examples in the README: https://github.com/drcandacemakedamoore/cleanX/blame/main/README.md#L135
