### 2021-08-22 93cca3e
Update README to reflect the CLI changes
