### 2021-08-22 bcbd6cb
Split show_images_in_df into two functions
