### 2021-08-22 04c938a
Add test for find_close_images
