### 2021-08-22 f6a28fd
[fix] Try to do shell expansion. May be helpful on Windows
