### 2021-08-23 drcandacemakedamoore/cleanX#35 @drcandacemakedamoore
I went through the notebooks and made all the paths relative. I hope this is easier for users to understand.
