### 2021-08-23 drcandacemakedamoore/cleanX#36 @drcandacemakedamoore
In the new version, we have updated the readme file to show what is supported in a fuller way. I am therefore closing this issue.
