### 2021-09-10 #3632 @anki-xyz
@cMadan can we maybe schedule a five min zoom or something? I feel stupid to not actually can do the things I am supposed to do... I think I need some kind of help. Thank you!
