### 2021-09-30 #3632 @henrykironde
Here are some recommendations @drcandacemakedamoore, feel free to let me know what is appropriate to add to your software.

1. Combine the `Api source docs` with other related documentation like `Medical professional documentation`, include Most of the sections in the README or all of it.
2. Add developer docs. Contributors should be able to set up this development environment with minimal effort and run tests locally. Example for reference https://retriever.readthedocs.io/en/latest/developer.html
3. Add an`environment.yml` to make it simple for users to set up the Conda virtual environment in the installation part.
3.  Move all the badges to the top. This makes it easy for users to notice the state of the software.
4. Explain your workflow in the developer docs. (2-3 sentences)
5. Since you are changing the main only at release, I strongly recommend that you create a branch called `development` for contributors to add patches before the release. Currently, you have several branches and it is hard to see any progress on the software after the release. Why the name `development` or dev? This is very clear for most users.
