### 2021-10-01 9d39716
Add code and CI job to generate conda environment files
