+++ conda-pkg/env.yml.in
@@ -1,0 +1,6 @@
+name: $env_name
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - $conda_recs
