### 2021-10-01 4b70a9a
Reduce verbosity in conda installs
