### 2021-10-01 #3632 @wvxvw
@henrykironde Can you please look at this branch: https://github.com/drcandacemakedamoore/cleanX/tree/wvxvw/review-fixes and tell me if it helps you with `conda` environments?
