### 2021-10-01 #3632 @henrykironde
Thanks for the Updates.

> I started working on this, and then I realized that... well, we build the package for number of platforms (you can see them in the readme), the dependencies are different for each platform

Could you explain to how the dependencies are different for each platform.

> Can you please look at this branch:
 I think that does not help me as a contributor to install from source. I would expect to see list of non python prerequisite tools. Once all are installed, it would be a clean `pip install `
