### 2021-10-02 #3632 @wvxvw
We don't use `pip` during development. So, if you want to use it, I could try to help you, but this might be hard... The way it would work is that you'd first build the package (we don't build source packages ourselves, so, better go with Wheel, because that's something we tried). Then, you'd install the built packages. I.e. the process is:

``` sh
python ./setup.py bdist_wheel
python -m pip install ./dist/*.whl
```

Make sure you have `wheel` package installed before you try to `bdist_wheel`.

The way to work on the project is to build the sources and then to install the resulting package. If you are using Anaconda, then you would use `conda build -c conda-forge ./conda-pkg/` after you have generated the necessary configuration files. An then `conda install -yqn $build_environment --use-local cleanx`.

---

I know it's a popular practice in Python community to do something like `pip install -e .` in the project directory... but, years of experience taught me that while this works OK for projects developed internally and only worked on by a small group of programmers inside a single organization, once projects are made public, this no longer works due to the difference in individual setups.

This also goes to answering your question about dependencies. It... depends. Depends on what you count. The way I've generated the environment files now, they only list the essential dependencies, similar to how the build configuration does. So, you may reasonably expect them to be the same... or at least very similar. However, another, and more popular way to generate the dependency list is to run `conda env export`.

The later will ensure that the dependencies you have are *exactly* like mine (whereas in the former case, you might get newer packages installed, because that's how `conda` constraint solver works). However, `conda` lists as dependencies platform libraries like `ld_impl_linux-64=2.36.1=hea4e1c9_0`. If you are on a Mac, there's no way you can reproduce this setup, because it specifically wants a library built for Linux. But, it's also not necessary for you to have it on a Mac, because if you start installing from the bare minimum, the constraint solver will never even try to installing this library (because it's conditioned on operating system).

So, really, if you want to build from source, please follow what the CI script does for the platform that you have. If you find it difficult to do so, please tell me the combination of Python version, Python distribution and operating system (I assume MacOS) you want to build on, and I'll extract the CI script into something that fits your platform.
