### 2021-10-02 #3632 @wvxvw
Actually, I just looked into environment files for Mac:

Python 3.9:

```yaml
name: cleanx-env-darwin-py39
channels:
  - conda-forge
  - defaults
dependencies:
  - pandas=1.3.3=py39h4d6be9b_0
  - numpy=1.21.2=py39h7eed0ac_0
  - matplotlib=3.4.3=py39h6e9494a_1
  - pillow=8.3.2=py39he9bb72f_0
  - tesserocr=2.5.1=py39h81cd7d3_2
  - libopencv=4.5.3=py39h84599ab_2
  - pytest=6.2.5=py39h6e9494a_0
  - pycodestyle=2.7.0=pyhd8ed1ab_0
  - sphinx=4.2.0=pyh6c4a22f_0
```

Python 3.8

```yaml
name: cleanx-env-darwin-py38
channels:
  - conda-forge
  - defaults
dependencies:
  - pandas=1.3.3=py38ha53d530_0
  - numpy=1.21.2=py38h49b9922_0
  - matplotlib=3.4.3=py38h50d1736_1
  - pillow=8.2.0=py38h83525de_1
  - tesserocr=2.5.1=py38h0a5c65b_2
  - libopencv=4.5.0=py38_3
  - pytest=6.2.5=py38h50d1736_0
  - pycodestyle=2.7.0=pyhd8ed1ab_0
  - sphinx=4.2.0=pyh6c4a22f_0
```

So, as you can see, they are *totally* different because of Python version. It's not just a tiny difference. These are entirely different packages.
