### 2021-10-02 #3632 @henrykironde
Thanks for the explanation @wvxvw, I still don't see the difference. Could you clarify that for me, please?

Also, I have looked through the development docs of core packages that you are using, and they use pip install.
pandas https://github.com/pandas-dev/pandas
numpy https://numpy.org/devdocs/dev/development_environment.html#other-build-options
matplotlib https://matplotlib.org/stable/users/installing.html
pillow https://pillow.readthedocs.io/en/latest/installation.html#basic-installation
I have personally installed these packages locally at some time when debugging some of the software projects I maintain.
Please, compare your installation guide to what your dependency packages have, and come up with some docs that enable developers to utilize your tool with ease.
