### 2021-10-02 #3632 @wvxvw
> I still don't see the difference. Could you clarify that for me, please?

Look at the part that follows the second equals sign, it tells you the package was built for a different version of Python. Every minor version of Python has different ABI. In other words, if you link against `libpython`, you must link against a specific version of it. You cannot have a version of a package that can be loaded by either Python 3.8 or 3.9, if it has a shared library that needs to link with `libpython`. Not only that, Python runtime reserves the right to change bytecode format between minor version, that's why Wheel format requires that you specify the Python's minor version in the name of the Wheel, if you are packaging bytecompiled Python files or shared libraries linked against `libpython`.

> Also, I have looked through the development docs of core packages that you are using, and they use pip install.

How they manage their development process is up to them... I believe, you, of course, realize that if you use `pip install` to install your package, this doesn't help you to deal with Anaconda's packages. So, if you want to develop for Anaconda, you need a different toolchain.

> Please, compare your installation guide to what your dependency packages have, and come up with some docs that enable developers to utilize your tool with ease.

At the moment we don't have an installation guide for developers. We only instruct the end users on how to install the package.

I will certainly add instructions for developers to the readme, but the ease is in the eye of the beholder. I'm still waiting for your response on the branch I've sent you. Does it help you? If not, why, and what can be improved? Please help me understand what parts don't you find easy.

Another question, that I would really like you to answer to help me understand your difficulties is whether you want to work on our project in the same way we work on it, or do you want us to make your working style possible with our project. I.e. I understand that your working style relies on `pip install -e .`, if you develop for PyPI and on `conda create -f env.yaml` if you develop for Anaconda. Am I correct about that?

Furthermore, we don't have any formal or informal process for accepting contributions from developers who aren't on our team. We are too few, and didn't anticipate the need for such a thing. So, we don't have any *requirements* for developers, and I believe the developers may use whatever tools they feel most convenient with to work on the project. I'm not sure that hand-holding them to create their working environment is my place.
