### 2021-10-02 #3632 @henrykironde
> Another question, that I would really like you to answer to help me understand your difficulties is whether you want to work on our project in the same way we work on it, or do you want us to make your working style possible with our project. I.e. I understand that your working style relies on pip install -e ., if you develop for PyPI and on conda create -f env.yaml if you develop for Anaconda. Am I correct about that?

What I am trying to say is that there are standards in software development, these standards need to be maintained by all users for the ecosystem to have fewer dependency problems. I did not create these standards, and that is why I pointed you to the docs of the packages you are using, to have some points of reference. This will also avoid backward compatibility issues where a package does not depend on a version as Python 3.8 or 3.9 to work.

> cleanX package is a pure Python package
It should be able to run on the development tools in the core Python ecosystem, otherwise you can state the difference that it only runs in a Conda environment.

I think at this moment, I have provided enough recommendations to improve the utility of the software, and the quality to my expectations of the "industrial" standard. Feel free to make only the changes you think are appropriate. I will review the software when you feel ready.
