### 2021-10-02 #3632 @wvxvw
> This will also avoid backward compatibility issues where a package does not depend on a version as Python 3.8 or 3.9 to work.

I'm sorry... you misinterpreted what I wrote. Please, look again at the files generated for environments using different versions of Python. It's *impossible* to have the same set of dependencies because of how Python works. It's not my choice, or the choice of Pandas or NumPy or any other package.

The development instructions they (Pandas, NumPy etc.) give you are the instructions for *a* version of Python. Maybe the latest supported by the given package, or maybe the latest supported by the released version of the package... they don't give you a way to set up your working environment for *all* versions.

If you want instructions for Python 3.9, MacOS, Anaconda Python, that is something I can give you. But, I cannot give you, for example, instructions on how to set up your environment if you want to work with Python 3.9, MacOS, python.org's Python, simply because we don't support this combination of platforms. I've never tried to use the package there, and have no access to the platform to even attempt to use it there.

You also misunderstand the backwards compatibility problem (in the context of our package). We are not developing for a particular version of Python, treating all older versions as backwards (in-)compatible. We develop for *different* versions of Python. From our perspective, they are all valid and valuable. As such, we don't have a "main" version, that aligns with a particular Python version, and "soon to be obsolete" versions that need less development effort and that soon will be abandoned.

> I think at this moment, I have provided enough recommendations to improve the utility of the software, and the quality to my expectations of the "industrial" standard.

Then I will treat the bullet point 3 in your initial post as being addressed by the branch `wvxvw/review-fixes`. It still needs some polishing of documentation, but the main feature added there is my response to your request.  If you are not satisfied by the changes, please feel free to revisit it.
