### 2021-10-03 #3632 @wvxvw
Actually, the same branch also closes the bullet point 5 on your list.

As for bullet point 6: our development process doesn't require it, but I will add it anyways. If you are planning on making a contribution, please read the documentation added in `wvxvw/review-fixes`.
