+++ binder/requirements.txt
@@ -4,4 +4,5 @@ numpy
 matplotlib
 pandas
 opencv-python-headless
+tesserocr
 Pillow
