### 2021-10-03 9e29304
Add medical professional docs to the rest of the docs
