+++ _sources/index.rst.txt
@@ -35,6 +35,16 @@ Command-Line Interface Documentation
 
 :doc:`CLI <../cli>`
 
+Developer's Guide
+-----------------
+
+:doc:`Developer's Guide <../developers>`
+
+Medical Professionals Documentation
+-----------------------------------
+
+:doc:`Medical Professionals Documentation <../medical-professionals>`
+
 
 Indices and tables
 ==================
