+++ _sources/developers.rst.txt
@@ -1,0 +1,212 @@
+=================
+Developer's Guide
+=================
+
+Sending Your Work
+=================
+
+We accept pull requests made through GitHub.  Alternatively, you can
+send patch files in an email (please find developers' emails on their
+profile pages).  As is usual, we request that the changes be rebased
+on the branch they are to be integrated into.  We also request that
+
+.. code-block:: bash
+
+   python ./setup.py test
+   python ./setup.py lint
+
+succeed with your changes applied.  We'll try our best to attribute
+your work to you, however, you need to release your work under
+compatible license for us to be able to use it.
+
+.. warning::
+
+   We don't use `git-merge` command, and if your submission has merge
+   commits, we'll have to remove them.  This means that in such case
+   commit hashes will be different from those in your original
+   submission.
+
+Setting Up Development Environment
+==================================
+
+.. warning::
+
+   You may find this setup unusual.  Please read the `Rationale`_
+   section to satisfy your curiosity.
+
+There isn't *a* development environment.  You may need to create an
+environment for every combination of supported versions of operating
+system, Python distribution and Python version.  If you believe that
+your code is portable accross versions and platforms, then, of course,
+you don't need all the possible combinations.  Our CI should be able
+to pick up the most obvious incompatibilities between those.
+
+The Way We Do It
+^^^^^^^^^^^^^^^^
+
+Suppose you want to check your changes against Python 3.7 from
+python.org on Linux, then you would check out our sources using `git`
+from our `GitHub repo`_, and:
+
+.. code-block:: bash
+
+   python3.7 -m venv .venv37
+   . ./.venv3.7/bin/activate
+   python ./setup.py install_dev
+
+What this will do is quite a bit different from what
+`setuptools.install` command would normally do: It will create an Egg
+with `cleanX` sources, install it, then it will look for development
+dependencies, and install them too.  This is likely to take a lot longer
+than what a typical project setup would do.  Please, be patient.
+
+Similarly, if you want to develop using Anaconda Python, you would:
+
+.. code-block:: bash
+
+   conda create -n cleanx-37
+   conda activate -n cleanx-37
+   conda run -n cleanx-37 python ./setup.py install_dev
+
+Similar to above, this will create a virtual environment, build
+`conda` package, install it and then add development dependencies to
+what was installed.  This will likely take a *very* long time.
+Please, be patient.
+
+The Traditional Way
+^^^^^^^^^^^^^^^^^^^
+
+Regradless of the downsides of this approach, we try to support more
+common ways to work with Python projects.  It's a common practice to
+"install" a project during development by either using `pip install
+--editable` command, or by using `conda` environment files.
+
+We provide limited support for this approach.  For instance, if you
+want to work on the project using `pip`, you could try:
+
+.. code-block:: bash
+
+   python3.7 -m venv .venv37
+   . ./.venv/bin/activate
+   pip install -e .[dev]
+
+If you encounter problems with this approach related to runnig tests
+or generating documentation, please let us know.  Unfortunately,
+supporting this worklfow is a low-priority task for us, as internally
+we don't rely on this to work.
+
+Similarly, if you want to develop using `conda` environment files, we
+provide limited support for that.
+
+The environment files are generated using:
+
+.. code-block:: bash
+
+   conda run -n cleanx-build python setup.py anaconda_gen_env
+
+However, this means you already have to have all dependencies
+installed.  We run this script in CI and archive the environment files
+it produces.  To find such file suitable for your platform you will
+need to navigate to `GitHub Actions dashboard`_, and look in the
+artifacts section of the selected job for artifacts named
+`cleanx-env-$python_version-$os`.  Select the one that applies to your
[…]
