### 2021-11-03 #3632 @henrykironde
@cMadan thank you for the reminder. I have constantly checked on the progress of the software submitted however, I still can not get the source to install. Current error messages am getting :

```Python
Traceback (most recent call last):
  File "/opt/miniconda3/envs/cleanx/bin/conda-build", line 11, in <module>
    sys.exit(main())
  File "/opt/miniconda3/envs/cleanx/lib/python3.9/site-packages/conda_build/cli/main_build.py", line 481, in main
    execute(sys.argv[1:])
  File "/opt/miniconda3/envs/cleanx/lib/python3.9/site-packages/conda_build/cli/main_build.py", line 470, in execute
    outputs = api.build(args.recipe, post=args.post, test_run_post=args.test_run_post,
  File "/opt/miniconda3/envs/cleanx/lib/python3.9/site-packages/conda_build/api.py", line 184, in build
    raise ValueError('No valid recipes found for input: {}'.format(recipe_paths_or_metadata))
ValueError: No valid recipes found for input: ['/Users/henrysenyondo/Downloads/cleanX/conda-pkg']
Traceback (most recent call last):
  File "./setup.py", line 448, in <module>
    setup(
  File "/opt/miniconda3/lib/python3.8/site-packages/setuptools/__init__.py", line 161, in setup
    return distutils.core.setup(**attrs)
  File "/opt/miniconda3/lib/python3.8/distutils/core.py", line 148, in setup
    dist.run_commands()
  File "/opt/miniconda3/lib/python3.8/distutils/dist.py", line 966, in run_commands
    self.run_command(cmd)
  File "/opt/miniconda3/lib/python3.8/distutils/dist.py", line 985, in run_command
    cmd_obj.run()
  File "./setup.py", line 404, in run
    super().run()
  File "./setup.py", line 348, in run
    raise RuntimeError('Couldn\'t build {} package'.format(name))
RuntimeError: Couldn't build cleanX package
```

Additionally the whole documentation of the package needs improvement. The instructions are not very clear and need better organization.  I recommend the authors to go through the instructions provided and test them out. An example

```python
conda create -n cleanx
conda activate -n cleanx.  # will error
python ./setup.py install_dev
```

Another example:

> Whether python.org Python or Anaconda Python are supported, it means that version 3.7, 3.8 and 3.9 are supported. We know for certain that 3.6 is not supported, and there will be no support in the future.

I believe the author means: Supports 3.7+.
(_Simplicity is better and easy to read_)
