### 2021-11-03 31e7905
[fix] Documentation about Anaconda development workflow
