### 2021-11-03 #3632 @wvxvw
@henrykironde My apologies, the installation instructions were incomplete. I've updated the documentation to reflect that:

``` sh
conda create -n cleanx
conda activate cleanx
python ./setup.py genconda
python ./setup.py install_dev
```

This is the right sequence of commands. In addition, about `setup.py genconda`, this command generates necessary Anaconda package configuration files, unfortunately, they are specific to Python version. Unfortunately, right now, it's not possible to have multiple configurations concurrently (but it may be fixed in the future, if you really find this inconvenient). This means, that if you switch between different versions of Python, you will have to re-run this command.

Additionally, I've noticed that you are using MacOS and a combination of Python 3.8 and Python 3.9. To be honest, I didn't know that was even possible. I.e. you seem to be running `conda` command with Python 3.8, but the conda packages itself reside in Python 3.9 installation... As we don't try this in our ordinary development, I cannot guarantee that your setup will work. Please try using a single version of Python if at all possible.

On top of this, since you are using MacOS, some time ago we've discovered that in the combination of MacOS and Python 3.9 you will not be able to get OpenCV to work. I'm not 100% sure at the moment what the issue was, but, I believe, it's related to XCode and some `clang` flags. Since it looks like you have Python 3.8 available, I'd recommend using that instead.
