### 2021-11-03 #3632 @henrykironde
```python
UnsatisfiableError: The following specifications were found
to be incompatible with the existing python installation in your environment:

Specifications:

  - cleanx -> python==3.8

Your python: python=3.9

If python is on the left-most side of the chain, that's the version you've asked for.
When python appears to the right, that indicates that the thing on the left is somehow
not available for the python version you are constrained to. Note that conda will not
change your python version to a different minor version unless you explicitly specify
that.

Traceback (most recent call last):
  File "./setup.py", line 448, in <module>
    setup(
  File "/opt/miniconda3/lib/python3.8/site-packages/setuptools/__init__.py", line 161, in setup
    return distutils.core.setup(**attrs)
  File "/opt/miniconda3/lib/python3.8/distutils/core.py", line 148, in setup
    dist.run_commands()
  File "/opt/miniconda3/lib/python3.8/distutils/dist.py", line 966, in run_commands
    self.run_command(cmd)
  File "/opt/miniconda3/lib/python3.8/distutils/dist.py", line 985, in run_command
    cmd_obj.run()
  File "./setup.py", line 404, in run
    super().run()
  File "./setup.py", line 360, in run
    raise RuntimeError('Couldn\'t install {} package'.format(name))
RuntimeError: Couldn't install cleanX package
````
