### 2021-11-03 #3632 @henrykironde
Somethings that I noted, like you said, we should not have two references for python 3.8 and 3.9. I guess we should try to create the env stating the python version to run the package. That may probably resolve some issue. Again thanks for the comments.
