### 2021-11-03 #3632 @drcandacemakedamoore
Is it possible that you downloaded Anaconda for Python 3.8, then upgraded
to Anaconda (which might give you one for Python 3.9)? Anaconda can
apparently sometimes upgrade itself without you asking it to (we just
discovered this today, and it actually gave us the same problem). If so I
think you could downgrade, or do a clean install of Anaconda. As we now
have the same problem, I expect Oleg will write more details on how we can
fix it.

On Wed, Nov 3, 2021 at 8:19 PM henry senyondo ***@***.***>
wrote:

> Somethings that I noted, like you said, we should not have two references
> for python 3.8 and 3.9. I guess we should try to create the env stating the
> python version to run the package. That may probably resolve some issue.
> Again thanks for the comments.
>
