### 2021-11-03 5b462c6
[fix] Prevent conda from randomly updating Python version
