### 2021-11-03 #3632 @wvxvw
Fix is on the way to `develop`.

If you are interested to know what happened, here's a synopsis:

`conda install xxx` will ignore Python version you configured for your virtual environment, and will install arbitrary version it likes, and if it means that in order to install this version `conda` needs to install a different version of Python, or a different version of `conda` etc, then so be it.

It didn't always work like that. But it did at the time of Python 2.7. Then, at some point the behavior improved, but at some later point the bad behavior returned. In the latest version of `conda`, the behavior described above is what you get.

So, why it worked just a week ago and why it doesn't work today: in order to build our package we install `conda-build` and `conda-verify`. We don't require a specific version of either of those packages as it's irrelevant to us. However, even though you probably created a virtual environment using Python 3.8, `conda` discovered `conda-build=3.21.5=py39h06a4308_0` (or similar, if you are on a Mac / Windows). Even though there's `conda-build-3.21.5=py38h06a4308_0`, `conda` decided that it's better to update everything in your environment... including Python and `conda` itself. It must be the case that `conda-build` was unavailable for Python 3.9, and so the newer version existed either solely in 3.7 or maybe `conda` in 3.7 wouldn't update Python and the latest version of `conda-build` was from 3.8.

In our build, we request specifically to build our package for the version of Python that was installed in your environment, i.e. 3.8. Because build happens in a separate environment (I don't know how to circumvent this, it's completely unnecessary and slows the build by a lot), the package was built successfully, but for the wrong Python version.

Worse yet, `conda` ignored Python requirement when testing the package (I've also tried to add it explicitly to the `test` section of the `meta.yaml` -- it has no effect. And that's how you ended up with the error you have.
