### 2021-11-04 74387c1
new image quality functions
