### 2021-11-04 9bb14bb
new image quality functions
