+++ test/test_cleanx.py
@@ -441,12 +441,12 @@ def test_show_close_images(monkeypatch):
     monkeypatch.setattr(iwork.image_functions.plt, 'show', lambda: None)
     iwork.show_close_images(image_directory, 2, 190)
 
-def test_image_to_histo(image):
+def test_image_to_histo():
     outpic = cv2.imread('testtocrop.jpg')
     hist = iwork.image_to_histo(outpic)
     assert len(hist) > 0
 
-def black_end_ratio(image_array):
+def black_end_ratio():
     outpic = cv2.imread('testtocrop.jpg')
     hist = iwork.black_end_ratio(outpic)
     assert hist > 0
