### 2021-11-05 f663e3b
image_functions fix histofunct
