### 2021-11-05 8f35a1f
update __init__.py with new functions
