+++ cleanX/image_work/__init__.py
@@ -48,6 +48,8 @@ from .image_functions import (
     image_quality_by_size,
     show_close_images,
     find_close_images,
+    image_to_histo,
+    black_end_ratio
 
     Rotator,
 )
