### 2021-11-05 3786364
update __init__.py with new functions
