+++ test/test_cleanx.py
@@ -444,10 +444,10 @@ def test_show_close_images(monkeypatch):
 def test_image_to_histo():
     outpic = cv2.imread('testtocrop.jpg')
     hist = iwork.image_to_histo(outpic)
-    assert len(hist) > 0
+    assert type(hist) > np.ndarray
 
 def black_end_ratio():
     outpic = cv2.imread('testtocrop.jpg')
-    hist = iwork.black_end_ratio(outpic)
-    assert hist > 0
+    ratio = iwork.black_end_ratio(outpic)
+    assert ratio > 0
 
