### 2021-11-05 7b0f776
fix testing output, by temporarily not testing image_to_hist
