+++ README.md
@@ -1,3 +1,6 @@
+
+# cleanX
+
 <p align="center">
     <img style="width: 30%; height: 30%"
         src="https://github.com/drcandacemakedamoore/cleanX/blob/main/test/cleanXpic.png">
@@ -25,19 +28,25 @@ radiological images.  JPEG files can be extracted from
 [![PYPI Version](https://img.shields.io/pypi/v/cleanX.svg)](https://pypi.org/project/cleanX)
 [![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/version.svg)](https://anaconda.org/doctormakeda/cleanx)
 
-Primary author: Candace Makeda H. Moore
-
-Other authors and contributors: Oleg Sivokon, Andrew Murphy
+Primary authors: Candace Makeda H. Moore, Oleg Sivokon, Andrew Murphy
 
 ### Documentation
 [![Documentation](https://img.shields.io/static/v1?label=docs&message=release&color=green)](https://drcandacemakedamoore.github.io/cleanX)
 
-# cleanX
+Online documentation at https://drcandacemakedamoore.github.io/cleanX/
 
-CleanX is an open source python library for exploring, cleaning and
-augmenting large datasets of X-rays, or certain other types of
-radiological images.  JPEG files can be extracted from
-[DICOM](https://www.dicomstandard.org/) files or used directly.
+You can also build up-to-date documentation by command.
+
+Documentation can be generated by command:
+
+python setup.py apidoc
+python setup.py build_sphinx
+
+The documentation will be generated in ./build/sphinx/html directory. Documentation is generated automatically as new functions are added.
+
+Special additional documentation for medical professionals with limited programming ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
+
+To get a high level overview of some of the functionality of the program you can look at the Jupyter notebooks inside workflow_demo.
 
 ## Requirements
 
@@ -298,19 +307,19 @@ program you can look at the Jupyter notebooks inside workflow_demo.
 # About using this library
 
 If you use the library, please cite the package.
-Only free to use this library according to license.
+CleanX is free ONLY when used according to license.
 
 You can get in touch with me by starting a [discussion](https://github.com/drcandacemakedamoore/cleanX/discussions/37) if you
 have a legitimate reason to use my library without open-sourcing your
-code base, or following other conditions, and I can make you
-specifically a different license.
+code base, or following other conditions, and I can make you,
+specifically, a different license.
 
 We are adding new functions and classes all the time. Many unit tests
 are available in the test folder. Test coverage is currently
 partial. Some newly added functions allow for rapid automated data
-augmentation (in ways that are realistic for radiological data). Some
-other classes and functions are for cleaning datasets including ones
-that:
+augmentation (in ways that are realistic for radiological data) and some
+preliminary image quality checks. Some other classes and functions are for
+cleaning datasets including ones that:
 
 * Get image and metadata out of dcm (DICOM) files into jpeg and csv
   files
