+++ README.md
@@ -1,10 +1,8 @@
-
-# cleanX
-
 <p align="center">
     <img style="width: 30%; height: 30%"
         src="https://github.com/drcandacemakedamoore/cleanX/blob/main/test/cleanXpic.png">
 </p>
+# cleanX
 
 [![Zenodo DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4725904.svg)](https://doi.org/10.5281/zenodo.4725904)
 [![License GPL-3](https://img.shields.io/github/license/drcandacemakedamoore/cleanX.svg)](https://raw.githubusercontent.com/drcandacemakedamoore/cleanX/main/LICENSE)
@@ -39,8 +37,8 @@ You can also build up-to-date documentation by command.
 
 Documentation can be generated by command:
 
-python setup.py apidoc
-python setup.py build_sphinx
+    python setup.py apidoc    
+    python setup.py build_sphinx
 
 The documentation will be generated in ./build/sphinx/html directory. Documentation is generated automatically as new functions are added.
 
