+++ README.md
@@ -14,14 +14,15 @@
 [![Anaconda-Server Badge](https://anaconda.org/doctormakeda/cleanx/badges/version.svg)](https://anaconda.org/doctormakeda/cleanx)
 [![Sanity](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-commit.yml/badge.svg)](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-commit.yml)
 [![Sanity](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-tag.yml/badge.svg)](https://github.com/drcandacemakedamoore/cleanX/actions/workflows/on-tag.yml)
+[![Documentation](https://img.shields.io/static/v1?label=docs&message=release&color=green)](https://drcandacemakedamoore.github.io/cleanX)
 
 
-CleanX is an open source python library for exploring, cleaning and
-augmenting large datasets of X-rays, or certain other types of
-radiological images.  JPEG files can be extracted from
-[DICOM](https://www.dicomstandard.org/) files or used directly.
 
-Primary authors: Candace Makeda H. Moore, Oleg Sivokon, Andrew Murphy
+CleanX is an open source python library for exploring, cleaning and augmenting
+large datasets of X-rays, or certain other types of radiological images.  JPEG
+files can be extracted from [DICOM](https://www.dicomstandard.org/) files or
+used directly.The primary authors are Candace Makeda H. Moore, Oleg Sivokon,
+and Andrew Murphy.
 
 
 ### The latest official release:
@@ -50,8 +51,6 @@ Documentation can be generated by command:
 The documentation will be generated in ./build/sphinx/html directory. Documentation is generated automatically as new functions are added.
 
 Special additional documentation for medical professionals with limited programming ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
-
-To get a high level overview of some of the functionality of the program you can look at the Jupyter notebooks inside workflow_demo.
 
 ## Requirements
 
