+++ README.md
@@ -27,17 +27,15 @@ directly. The primary authors are Candace Makeda H. Moore, Oleg Sivokon, and And
 ### Documentation
 
 Online documentation at https://drcandacemakedamoore.github.io/cleanX/
+You can also build up-to-date documentation by command as follows:
 
-You can also build up-to-date documentation by command.
-
-Documentation can be generated by command:
 
     python setup.py apidoc    
     python setup.py build_sphinx
 
-The documentation will be generated in ./build/sphinx/html directory. Documentation is generated automatically as new functions are added.
+The documentation will be generated in ./build/sphinx/html directory. 
 
-Special additional documentation for medical professionals with limited programming ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
+Special additional documentation for medical professionals with limited programming ability is available on the wiki (https://github.com/drcandacemakedamoore/cleanX/wiki/Legacy-medical-professional-documentation). To get a high level overview of some of the functionality of the program you can look at the Jupyter notebooks inside workflow_demo folder.
 
 ## Requirements
 
@@ -296,13 +294,6 @@ python setup.py build_sphinx
 The documentation will be generated in `./build/sphinx/html`
 directory. Documentation is generated automatically as new functions
 are added.
-
-Special additional documentation for medical professionals with
-limited programming ability is available on the wiki
-(https://github.com/drcandacemakedamoore/cleanX/wiki/Medical-professional-documentation).
-
-To get a high level overview of some of the functionality of the
-program you can look at the Jupyter notebooks inside workflow_demo.
 
 
 ## About using this library
