### 2021-11-06 581157a
more readme fixes
