+++ README.md
@@ -283,7 +283,7 @@ our CI process.
 
 ### Build up-to-date documentation
 
-Documentation can be generated by command:
+Documentation can be [generated by command](#documentation).
 
 ``` sh
 python setup.py apidoc
