+++ README.md
@@ -283,16 +283,9 @@ our CI process.
 
 ### Build up-to-date documentation
 
-Documentation can be [generated by command](#documentation).
-
-``` sh
-python setup.py apidoc
-python setup.py build_sphinx
-```
-
-The documentation will be generated in `./build/sphinx/html`
-directory. Documentation is generated automatically as new functions
-are added.
+Documentation can be [generated by command](#documentation). The documentation 
+will be generated in a `./build/sphinx/html` directory. Documentation is generated
+automatically as new functions are added.
 
 
 ## About using this library
