+++ cleanX/image_work/steps.py
@@ -192,6 +192,41 @@ class Crop(Step):
             logging.exception(e)
             return None, e
 
+# class Tee(Step):
+#     """This step makes duplicate images, then opens two parallel pipelines"""
+        # Unfinished
+#     def apply(self, image_data):
+
+class Salt(Step):
+    """This class takes the image and applies the salting function
+    (augments with noise). In present version reccomended to run on copies.
+    In future versions can be run after a Tee step"""
+
+    def __init__(
+        self,
+        kernel=(5,5),
+        erosion_interations=90,
+        dilation_iterations=10,
+        cache_dir=None,
+    ):
+        super().__init__(cache_dir)
+        self.kernel = kernel
+        self.erosion_iterations = erosion_interations
+        self.dilation_iterations = dilation_iterations
+
+    def apply(self, image_data):
+        erosion = cv2.erode(
+            image_data,
+            self.kernel,
+            iterations=self.erosion_iterations,
+        )
+        dilation = cv2.dilate(
+            erosion,
+            self.kernel,
+            iterations=self.dilation_iterations,
+        )
+        salty_noised = (image_data + (erosion - dilation))
+        return salty_noised
 
 class Normalize(Step):
     """This class makes a simple normalizing to get values 0 to 255."""
