### 2021-11-14 afca5a3
fix salt with returning 2 values
