### 2021-11-15 b26c536
more image augmenting steps for pipeline
