### 2021-11-15 965e8c1
remove salt function
