### 2021-11-15 c073ef1
remove salt function
