### 2021-11-19 0306406
PSG pipeline only the PSG, no CLI, not persistence
