+++ cleanX/image_work/steps.py
@@ -4,6 +4,8 @@ import logging
 import os
 import json
 import inspect
+from multiprocessing import Queue
+from queue import Empty
 
 import numpy as np
 import cv2
@@ -37,8 +39,9 @@ class Step(metaclass=RegisteredStep):
         """
         self.cache_dir = cache_dir
         self.position = None
+        self.transaction_started = False
 
-    def apply(self, image_data):
+    def apply(self, image_data, image_name):
         """
         This is the method that will be called to do the actual image
         transformation.  This function must not raise exceptions, as it
@@ -109,6 +112,12 @@ class Step(metaclass=RegisteredStep):
             logging.exception(e)
             return e
 
+    def begin_transaction(self):
+        self.transaction_started = True
+
+    def commit_transaction(self):
+        self.transaction_started = False
+
     def __reduce__(self):
         return self.__class__, (self.cache_dir,)
 
@@ -127,6 +136,71 @@ class Step(metaclass=RegisteredStep):
     def from_cmd_args(cls, cmd_args):
         evaled_args = eval('dict({})'.format(cmd_args))
         return cls(**evaled_args)
+
+
+class Aggregate(Step):
+
+    def __init__(self, pre, agg, post, cache_dir=None):
+        super().__init__(cache_dir=cache_dir)
+        self.pre = pre
+        self.agg = agg
+        self.post = post
+        self.accumulator = None
+
+    def begin_transaction(self):
+        super().begin_transaction()
+        self.accumulator = self.pre()
+
+    def commit_transaction(self):
+        super().commit_transaction()
+        super().write(*self.post(self.accumulator))
+
+    def aggregate(self, accum, new):
+        try:
+            self.accumulator = self.agg(accum[0], accum[1], new[0], new[1])
+        except Exception as e:
+            return None, e
+        return self.accumulator, None
+
+    def __reduce__(self):
+        return self.__class__, (
+            self.pre,
+            self.agg,
+            self.post,
+            self.cache_dir,
+        )
+
+
+class Mean(Aggregate):
+
+    def __init__(self, cache_dir=None):
+        super().__init__(
+            pre=self.do_pre,
+            agg=self.do_agg,
+            post=self.do_post,
+            cache_dir=cache_dir,
+        )
+
+    def do_pre(self):
+        return None, None
+
+    def do_agg(self, acc_data, acc_name, image_data, image_name):
+        if acc_data is None:
+            return image_data, [image_name]
+        acc_name.append(image_name)
+        accw, acch = acc_data.shape[:2]
+        iw, ih = image_data.shape[:2]
+        mw = max(accw, iw)
+        mh = max(acch, ih)
+        acc_data = np.float32(cv2.resize(acc_data, (mw, mh)))
+        image_data = np.float32(cv2.resize(image_data, (mw, mh)))
+        return acc_data + image_data, acc_name
+
+    def do_post(self, acc):
+        return np.uint8(acc[0] / len(acc[1])), acc[1][-1]
+
+    def __reduce__(self):
+        return self.__class__, (self.cache_dir,)
 
 
 class Acquire(Step):
@@ -158,6 +232,7 @@ class Save(Step):
             self.extension,
         )
         try:
+            print("exporting", path)
             cv2.imwrite(
                 os.path.join(self.target, name),
                 image_data,
@@ -177,13 +252,13 @@ class Save(Step):
 class Crop(Step):
     """This class crops image arrays of black frames"""
 
-    def apply(self, image_data):
+    def apply(self, image_data, image_name):
 
         try:
             nonzero = np.nonzero(image_data)
             y_nonzero = nonzero[0]
[…]
