### 2021-11-21 c34080f
[fix] Pipleine tests due to changed interface
