### 2021-11-26 192c060
Simplify aggregate functionality
