### 2021-11-27 6c5d23b
change iwork init to include Aggregate
