+++ cleanX/image_work/__init__.py
@@ -70,6 +70,7 @@ from .steps import (
     HistogramNormalize,
     Sharpie,
     BlurEdges,
+    Aggregate,
     Mean,
 )
 
