### 2021-11-27 a43357b
add a group histo function
