### 2021-11-28 3252082
add test for additional histo
