### 2021-11-28 4063975
fix show_images function by alt function
