### 2021-11-28 428cb35
fix show my images in df function
