### 2021-11-28 1359483
fix show_images_in_df function
