### 2021-12-02 3ac7989
new function for segmentation
