### 2021-12-02 27d659c
new function for segmentation,linted
