### 2021-12-02 f8cb2d0
new function for segmentation, relinted times 2
