### 2021-12-02 c0e0cd5
new function for segmentation documentation
