### 2021-12-02 e8222b1
improve segment function and test
