+++ cleanX/image_work/image_functions.py
@@ -1857,7 +1857,7 @@ def black_end_ratio(image_array):
     return ratio
 
 
-def outline_segment_by_otsu(image_to_transform):
+def outline_segment_by_otsu(image_to_transform, blur_k_size=1):
     """
     This is a function to turn an Xray into an outline
     with a specific method that involves an implementation
@@ -1865,9 +1865,14 @@ def outline_segment_by_otsu(image_to_transform):
     the result is line images that can be very useful 
     in and of themselves to run a nueral net on
     or can be used for segmentation in some cases
+    blur_k_size used in a blur to make ourlines less detailed
+    if set to a higher value, 0 < values < 100, and odd
 
     :param image_to_transform: the image name
     :type image_to_transform: string
+    :param blur_k_size: must be odd and value <100, kernel to blur 
+    to make ourlines less detailed
+    :type blur_k_size: int
 
     :return: edges (an image with lines)
     :rtype: numpy.ndarray
@@ -1900,6 +1905,12 @@ def outline_segment_by_otsu(image_to_transform):
             if output_image[x, y] < thresh:
                 # lets set this to zero
                 output_image[x, y] = 0
+    if blur_k_size % 2 !=0:
+    # blur
+        output_image = cv2.medianBlur(output_image,blur_k_size)
+    else:    
+        output_image = cv2.medianBlur(output_image,(blur_k_size + 1))
+               
     # now use canny to get edges
     edges = cv2.Canny(output_image, 50, 230)
     return edges
