### 2021-12-02 d5abef0
new function for segmentation remade, linting, testing
