### 2021-12-02 ecd24da
improve segment function  tests
