+++ test/test_cleanx.py
@@ -448,11 +448,13 @@ def test_show_close_images(monkeypatch):
 #     assert type(hist) is np.ndarray
 
 def test_black_end_ratio():
-    outpic = cv2.imread('testtocrop.jpg')
+    vovo = os.path.join(image_directory, 'testtocrop.jpg')
+    outpic = cv2.imread(vovo)
     ratio = iwork.black_end_ratio(outpic)
-    assert type(ratio) is 0
+    assert type(ratio) is float
 
 def test_outline_segment_by_otsu():
-    outpic = 'testtocrop.jpg'
-    outlined = iwork.test_outline_segment_by_otsu(outpic,11)
+    vovo = os.path.join(image_directory, 'testtocrop.jpg')
+    #outpic = 'testtocrop.jpg'
+    outlined = iwork.outline_segment_by_otsu(vovo,11)
     assert type(outlined) is np.ndarray 
