### 2021-12-02 10234a0
Enable easier local testing and lintin
