+++ conda-pkg/meta.yaml.in
@@ -3,7 +3,7 @@ package:
   version: ${version}
 
 source:
-  path: ..
+  git_url: ..
 
 requirements:
   host:
