### 2021-12-03 a23314f
Add some documentation about local testing and building
