### 2021-12-03 0e4e6dd
add binarize function and test
