+++ cleanX/image_work/__init__.py
@@ -51,6 +51,7 @@ from .image_functions import (
     image_to_histo,
     black_end_ratio,
     outline_segment_by_otsu,
+    binarize_by_otsu,
 
     Rotator,
 )
