+++ test/test_cleanx.py
@@ -458,3 +458,9 @@ def test_outline_segment_by_otsu():
     #outpic = 'testtocrop.jpg'
     outlined = iwork.outline_segment_by_otsu(vovo,11)
     assert type(outlined) is np.ndarray 
+
+def binarize_by_otsu():
+    vovo = os.path.join(image_directory, 'testtocrop.jpg')
+    #outpic = 'testtocrop.jpg'
+    outlined = iwork.binarize_by_otsu(vovo,11)
+    assert type(outlined) is np.ndarray 
