### 2021-12-04 0328a3a
make one otsu more efficient
