+++ cleanX/image_work/__init__.py
@@ -52,6 +52,7 @@ from .image_functions import (
     black_end_ratio,
     outline_segment_by_otsu,
     binarize_by_otsu,
+    column_sum_folder,
 
     Rotator,
 )
