+++ cleanX/image_work/image_functions.py
@@ -1896,15 +1896,18 @@ def outline_segment_by_otsu(image_to_transform, blur_k_size=1):
     index_of_max_val = np.argmax(inter_class_vari)
 
     thresh = bin_mids[:-1][index_of_max_val]
-    width, height = image_to_transform.shape[0], image_to_transform.shape[1]
+    # width, height = image_to_transform.shape[0], image_to_transform.shape[1]
     # this gave the thresh based on Otsu, now apply it
     output_image = image_to_transform
-    for x in range(width):
-        for y in range(height):
-            # for the given pixel at w,h, check value against the threshold
-            if output_image[x, y] < thresh:
-                # lets set this to zero
-                output_image[x, y] = 0
+    # for x in range(width):
+    #     for y in range(height):
+    #         # for the given pixel at w,h, check value against the threshold
+    #         if output_image[x, y] < thresh:
+    #             # lets set this to zero
+    #             output_image[x, y] = 0
+    mask = output_image < thresh
+    output_image[mask] = 0
+    # output_image[~mask] = 255
     if blur_k_size % 2 != 0:
         # blur
         output_image = cv2.medianBlur(output_image, blur_k_size)
