+++ cleanX/image_work/__init__.py
@@ -77,6 +77,7 @@ from .steps import (
     GroupHistoHtWt,
     GroupHistoPorportion,
     Mean,
+    OtsuBinarize,
 )
 
 
