### 2021-12-04 68d2390
add otsu lines as a step
