+++ cleanX/image_work/__init__.py
@@ -78,6 +78,7 @@ from .steps import (
     GroupHistoPorportion,
     Mean,
     OtsuBinarize,
+    OtsuLines,
 )
 
 
