+++ cleanX/image_work/__init__.py
@@ -79,6 +79,7 @@ from .steps import (
     Mean,
     OtsuBinarize,
     OtsuLines,
+    Projection,
 )
 
 
