### 2021-12-04 1801c25
[fix] Make sure not to delete cache directory while it's still needed
