### 2021-12-05 d8f0995
Parser for PSG pipeline description
