### 2021-12-11 c28f33e
[wip] Redesign pipeline to deal with multiple branches
