+++ cleanX/image_work/pipeline.py
@@ -6,11 +6,13 @@ import logging
 import multiprocessing
 
 from collections import defaultdict
-from concurrent.futures import ProcessPoolExecutor
+from multiprocessing.pool import Pool
 from tempfile import TemporaryDirectory
 from glob import glob
 
-from .steps import Aggregate, Step
+from .steps import Aggregate, Step, Serial, Parallel
+
+import numpy as np
 
 
 class MultiSource:
@@ -172,6 +174,9 @@ class PSG:
         deps = set(s.id() for s in self._deps[step.id()])
         return ids.intersection(deps)
 
+    def immediate_dependencies(self, step):
+        yield from self._deps[step.id()]
+
     def dependencies(self, goal):
         pairs = dict(self._deps)
         tier = pairs[goal.id()]
@@ -217,8 +222,6 @@ class Pipeline:
             self.batch_size = batch_size or os.cpu_count() or 1
 
         self.counter = 0
-
-        self.process_lock = multiprocessing.Lock()
 
     def workspace(self):
         """
@@ -270,111 +273,113 @@ class Pipeline:
                        for the images to be processed.
         :type source: Iterable
         """
-        with self.process_lock:
-            with self.workspace() as td:
-                previous_step = self.find_previous_step()
-                processed_step = None
-                deps = tuple(reversed(tuple(self.steps.dependencies(goal))))
-                dirty = []
-                for n, step in enumerate(deps):
-                    step.cache_dir = os.path.join(td, str(self.counter))
-                    os.mkdir(step.cache_dir)
-                    self.counter += 1
-                    self.begin_transaction(step)
+        with self.workspace() as td:
+            previous_step = self.find_previous_step()
+            processed_step = None
+            deps = tuple(reversed(tuple(self.steps.dependencies(goal))))
+            dirty = []
+            for n, step in enumerate(deps):
+                step.cache_dir = os.path.join(td, str(self.counter))
+                os.mkdir(step.cache_dir)
+                self.counter += 1
+                self.begin_transaction(step)
+                sdeps = tuple(self.steps.immediate_dependencies(step))
 
-                    if previous_step is None:
-                        srciter = enumerate(source)
-                    else:
-                        files = os.listdir(previous_step.cache_dir)
-                        srciter = (
-                            (i, os.path.join(previous_step.cache_dir, f))
-                            for i, f in enumerate(files)
+                if previous_step is None:
+                    srciter = ((s,) for s in source)
+                else:
+                    transposed, longest = [], 0
+                    for d in sdeps:
+                        abspaths = tuple(
+                            os.path.join(d.cache_dir, f)
+                            for f in os.listdir(d.cache_dir)
                         )
-                    self.process_step(step, srciter)
+                        longest = max(longest, len(abspaths))
+                        transposed.append(abspaths)
+                    for t in transposed:
+                        t.extend([None] * (longest - len(t)))
+                    scriter = np.array(transposed).T
+                self.process_step(step, srciter, sdeps)
 
-                    processed_step = previous_step
-                    previous_step = step
+                processed_step = previous_step
+                previous_step = step
 
-                    dirty.append(processed_step)
-                    dirty = self.cleanup(dirty, deps[n:])
-                    self.commit_transaction(step)
+                dirty.append(processed_step)
+                dirty = self.cleanup(dirty, deps[n:])
+                self.commit_transaction(step)
 
-                self.counter = 0
-                self.update_counter()
+            self.counter = 0
+            self.update_counter()
 
-    def process_step(self, step, srciter):
-        batch = []
-        for i, s in srciter:
-            logging.info('processing image: %s', s)
-            step.position = i
-            # TODO(wvxvw): Move reading into multiple processes.
-            sdata, err = step.read(s)
-            if err is not None:
[…]
