+++ cleanX/image_work/__init__.py
@@ -94,6 +94,7 @@ from .steps import (
     OtsuLines,
     Projection,
     CleanRotate,
+    Or,
 )
 
 

[…]
