### 2021-12-18 8f7318d
multi-rotate function without PIL
