### 2021-12-18 0a0daca
[fix] Maybe we can prevent conda from updating this way
