### 2021-12-18 ce4cb61
Path Wheel and fix the version of CI conda-related stuff
