### 2021-12-18 e6cd2cd
Add note about installing tesseract libraries
