+++ cleanX/image_work/image_functions.py
@@ -13,8 +13,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pandas as pd
 
-from PIL import Image
-from PIL import Image, ImageOps
+# from PIL import Image
+# from PIL import Image, ImageOps
 
 try:
     def __fix_tesserocr_locale():
@@ -91,27 +91,27 @@ def crop_np(image_array):
     ]
 
 
-def crop_pil(image):
-    """
-    Crops black edges of an Pil/Pillow image
+# def crop_pil(image):
+#     """
+#     Crops black edges of an Pil/Pillow image
 
-    :param image_array: Image.
-    :type image_array: image
+#     :param image_array: Image.
+#     :type image_array: image
 
 
-    :return: PIL Image with black margins cropped.
-    :rtype: PIL.Image
-    """
-    mode = image.mode
-    return Image.fromarray(
-        crop_np(np.array(image)),
-        mode=mode,
-    )
+#     :return: PIL Image with black margins cropped.
+#     :rtype: PIL.Image
+#     """
+#     mode = image.mode
+#     return Image.fromarray(
+#         crop_np(np.array(image)),
+#         mode=mode,
+#     )
 
 
 def crop(image):
     """
-    Crops an image of a black frame: does both PIL and opencv2 images
+    Crops an image of a black frame: does !both PIL and! opencv2 images
 
     :param image: Image
     :type image: This can be either a NumPy array holding image data,
@@ -120,8 +120,8 @@ def crop(image):
     :return: Image cropped of black edges
     :rtype: Same as input type.
     """
-    if isinstance(image, Image.Image):
-        return crop_pil(image)
+    # if isinstance(image, Image.Image):
+    #     return crop_pil(image)
     if isinstance(image, np.ndarray):
         return crop_np(image)
 
@@ -186,32 +186,32 @@ def salting(img):
     return salty_noised
 
 
-def simple_rotation_augmentation(angle_list1, image):
-    """
-    This function takes one picture and rotates is by a number
-    of degrees is angle_list1.This function can be used with the
-    augment_and_move function as follows (example):
+# def simple_rotation_augmentation(angle_list1, image):
+#     """
+#     This function takes one picture and rotates is by a number
+#     of degrees is angle_list1.This function can be used with the
+#     augment_and_move function as follows (example):
 
-    .. code-block:: python
+#     .. code-block:: python
 
-       augment_and_move(
-           'D:/my_academia/dataset/random_within_domain',
-           'D:/my_academia/elo',
-           [partial(simple_rotation_augmentation, 5)],
-       )
+#        augment_and_move(
+#            'D:/my_academia/dataset/random_within_domain',
+#            'D:/my_academia/elo',
+#            [partial(simple_rotation_augmentation, 5)],
+#        )
 
-    :param image: Image.
-    :type image: Image (JPEG)
+#     :param image: Image.
+#     :type image: Image (JPEG)
 
-    :return: rotated image
-    :rtype: PIL image
-    """
-    if isinstance(image, str):
-        image4R = Image.open(image)
-    else:
-        image4R = image
-    rotated1 = image4R.rotate(angle_list1)
-    return rotated1
+#     :return: rotated image
+#     :rtype: PIL image
+#     """
+#     if isinstance(image, str):
+#         image4R = Image.open(image)
+#     else:
+#         image4R = image
+#     rotated1 = image4R.rotate(angle_list1)
+#     return rotated1
 
 
 def simple_rotate_no_pil(image, angle, center=None, scale=1.0):
@@ -273,40 +273,40 @@ def blur_out_edges(image):
     return blurred_edge_image
 
 
-def reasonable_rotation_augmentation(angle1, angle2, number_slices, image):
-    """
[…]
