### 2021-12-18 ec84690
update readme to leave off PIL + linting
