### 2021-12-18 0008e33
[fix] Don't escape glob expression
