### 2021-12-18 fea5c2f
[wip] More robust timeline def using Lark
