### 2021-12-19 42c5153
[wip] Redesign pipeline to use new step graph
