+++ cleanX/image_work/pipeline.py
@@ -4,15 +4,22 @@ import os
 import shutil
 import logging
 import multiprocessing
+import time
 
 from collections import defaultdict
-from multiprocessing.pool import Pool
+from multiprocessing import Queue, Process
+from multiprocessing.pool import ThreadPool
 from tempfile import TemporaryDirectory
 from glob import glob
+from queue import Empty
 
 from .steps import Aggregate, Step, Serial, Parallel
 
 import numpy as np
+
+
+class Loop(ValueError):
+    pass
 
 
 class MultiSource:
@@ -112,17 +119,7 @@ class NoPathToGoal(LookupError):
 
 class PSG:
 
-    @classmethod
-    def create(cls, source):
-        if not source:
-            return cls(())
-        if isinstance(source[0], Step):
-            pairs = tuple(zip(source[1:], source))
-        else:
-            pairs = source
-        return cls(pairs)
-
-    def __init__(self, pairs):
+    def __init__(self, pipeline_def):
         # pairs:
         #   A -> B
         #   B -> C
@@ -154,43 +151,129 @@ class PSG:
         #   E, C, B, A, D
         # or;
         #   C, E, B, A, D
-        mapping = defaultdict(list)
-        nodes = {}
-        for a, b in pairs:
-            mapping[a.id()].append(b)
-            nodes[a.id()] = a
-            nodes[b.id()] = b
-        for a, b in pairs:
-            if b.id() not in mapping:
-                mapping[b.id()] = []
-        self._deps = dict(mapping)
-        self._nodes = nodes
+        self._def = pipeline_def
+        self._ready = {}
+        self._in_flight = set(())
 
-    def as_dict(self):
-        return self._nodes.items()
+    def goal(self):
+        return self._def.goal
 
-    def depends(self, step, others):
-        ids = set(o.id() for o in others)
-        deps = set(s.id() for s in self._deps[step.id()])
-        return ids.intersection(deps)
+    def deps(self, goal):
+        # linearized dependencies of the goal
+        found = set(())
+        unresolved = set(goal.variables)
+        while unresolved:
+            next_gen = set(())
+            next_unresolved = set(())
+            for u in unresolved:
+                s = self._def.steps[u]
+                if s not in found:
+                    next_gen.add(s)
+            for s in next_gen:
+                yield s
+                for v in s.variables:
+                    next_unresolved.add(v)
+            unresolved = next_unresolved
 
-    def immediate_dependencies(self, step):
-        yield from self._deps[step.id()]
+    def available(self, deps):
+        # selects dependencies whose preconditions have been met
+        for d in deps:
+            if d in self._in_flight:
+                continue
+            if not d.variables:
+                yield d
+            elif all(v in self._ready for v in d.variables):
+                yield d
 
-    def dependencies(self, goal):
-        pairs = dict(self._deps)
-        tier = pairs[goal.id()]
-        del pairs[goal.id()]
+    def resolve(self, dep, deps):
+        # mark dependency as resolved and return the unresolved ones
+        for v in self.out(dep):
+            self._ready[v] = dep
+        self._in_flight.pop(dep)
+        for d in deps:
+            if d != dep:
+                yield d
 
-        yield goal
+    def out(self, dep):
+        for v, s in self._def.steps.items():
+            if dep == s:
+                yield v
 
-        while tier:
-            next_tier = []
-            for t in tier:
-                next_tier += pairs[t.id()]
-                del pairs[t.id()]
-                yield t
-            tier = next_tier
+    def put(self, dep):
+        # remember this step as in-flight, so we don't mark it as
+        # available later.
+        self._in_flight.add(dep)
+
+    def pending(self, dep):
+        # True, if dep is either in-flight, or unresolved
[…]
