### 2021-12-20 b390cc1
Update minor version of setup-conda due to CVE
