+++ .github/workflows/on-commit.yml
@@ -91,7 +91,7 @@ jobs:
       - uses: actions/checkout@v2
         with:
           fetch-depth: 0
-      - uses: s-weigand/setup-conda@v1
+      - uses: s-weigand/setup-conda@v1.0.6
         with:
           conda-channels: conda-forge
           python-version: ${{ matrix.python-version }}
