### 2021-12-22 5d822ab
new function for total pixel
