### 2021-12-24 7e7b8fc
addd group pixelsum histos
