### 2021-12-24 596b73c
Let's see if explicitly requesting python in our environment helps
