+++ .github/workflows/on-tag.yml
@@ -56,7 +56,7 @@ jobs:
           conda-channels: conda-forge
           python-version: ${{ matrix.python-version }}
       - run: conda install conda-build anaconda-client conda-verify
-      - run: conda create --name cleanx-build
+      - run: conda create --name cleanx-build python=${{ matrix.python-version }}
       - run: >-
           conda run -n cleanx-build
           python setup.py genconda -t ${{ matrix.python-version }}
