+++ .github/workflows/on-commit.yml
@@ -95,7 +95,7 @@ jobs:
         with:
           conda-channels: conda-forge
           python-version: ${{ matrix.python-version }}
-      - run: conda create --name cleanx-build
+      - run: conda create --name cleanx-build python=${{ matrix.python-version }}
       - run: conda run -n cleanx-build python setup.py bdist_egg
       - run: conda run -n cleanx-build python setup.py genconda
         # conda-build installation and the build cannot be in an
