+++ cleanX/image_work/__init__.py
@@ -38,6 +38,7 @@ from .image_functions import (
     simple_spinning_template,
     make_contour_image,
     hist_sum_of_pixels_across_set,
+    crop_np_white,
     avg_image_maker,
     set_image_variability,
     find_outliers_sum_of_pixels_across_set,
