### 2021-12-25 0643a87
better overall crop, update affil on paper
