### 2021-12-28 0da37e7
image quality function and test
