### 2021-12-28 2ed313b
further update quality function
