### 2021-12-29 60aebcb
fix failing test on image quality function
