+++ cleanX/image_work/__init__.py
@@ -72,7 +72,8 @@ from .steps import (
     Step,
     Acquire,
     Save,
-    Crop,
+    BlackEdgeCrop,
+    WhiteEdgeCrop,
     Normalize,
     HistogramNormalize,
     Sharpie,
