### 2021-12-30 8b0faa8
delete extra checkpoint file plus minor changes to docu
