### 2021-12-31 6bc95f2
fourier funct and testing updates
