+++ test/test_pipeline.py
@@ -24,6 +24,7 @@ from cleanX.image_work import (
     WhiteEdgeCrop,
     Mean,
 )
+from cleanX.image_work.steps import FourierTransf
 # from cleanX.image_work.steps import BlackEdgeCrop, WhiteEdgeCrop
 
 
@@ -73,6 +74,20 @@ def test_crop_images():
             Acquire(),
             BlackEdgeCrop(),
             WhiteEdgeCrop(),
+            Save(td),
+        ))
+        p.process(src)
+        src_files = set(f for f in os.listdir(src_dir) if f.endswith('.jpg'))
+        dst_files = set(os.listdir(td))
+        assert src_files == dst_files
+
+def test_fourier_transf():
+    src_dir = image_directory
+    with TemporaryDirectory() as td:
+        src = DirectorySource(src_dir)
+        p = create_pipeline(steps=(
+            Acquire(),
+            FourierTransf(),
             Save(td),
         ))
         p.process(src)
