+++ cleanX/image_work/__init__.py
@@ -67,6 +67,7 @@ from .steps import (
     Step,
     Acquire,
     Save,
+    FourierTransf,
     BlackEdgeCrop,
     WhiteEdgeCrop,
     Normalize,
