### 2021-12-31 c80e1b6
projection step and test
