+++ cleanX/image_work/__init__.py
@@ -72,6 +72,7 @@ from .steps import (
     WhiteEdgeCrop,
     Normalize,
     HistogramNormalize,
+    ProjectionHorizoVert,
     Sharpie,
     BlurEdges,
     Aggregate,
