### 2022-01-01 74eac50
cut and pad functions and testing for them
