+++ test/test_cleanx.py
@@ -445,3 +445,15 @@ def test_fourier_transf():
     sample_imager = cv2.imread(sample_image)
     transformed = iwork.fourier_transf(sample_imager)
     assert sample_imager.shape[0] == transformed.shape[0]
+
+def test_pad_to_size():
+    sample_image = os.path.join(image_directory, 'testtocrop.jpg')
+    sample_imager = cv2.imread(sample_image)
+    padded = iwork.pad_to_size(sample_imager, 5000, 5000)
+    assert type(padded) is np.ndarray
+
+def test_cut_to_size():
+    sample_image = os.path.join(image_directory, 'testtocrop.jpg')
+    sample_imager = cv2.imread(sample_image)
+    padded = iwork.cut_to_size(sample_imager, 3, 2)
+    assert type(padded) is np.ndarray 
