### 2022-01-01 2bbfaab
add crop_or_cut function and testing
