+++ cleanX/image_work/__init__.py
@@ -54,6 +54,7 @@ from .image_functions import (
     fourier_transf,
     pad_to_size,
     cut_to_size,
+    cut_or_pad,
 
     Rotator,
 )
