+++ test/test_cleanx.py
@@ -457,3 +457,9 @@ def test_cut_to_size():
     sample_imager = cv2.imread(sample_image)
     padded = iwork.cut_to_size(sample_imager, 3, 2)
     assert type(padded) is np.ndarray 
+
+def test_cut_or_pad():
+    sample_image = os.path.join(image_directory, 'testtocrop.jpg')
+    sample_imager = cv2.imread(sample_image)
+    padded = iwork.cut_or_pad(sample_imager, 5000, 2)
+    assert type(padded) is np.ndarray 
