### 2022-01-01 3e059f1
better rotate function and test
