+++ test/test_pipeline.py
@@ -24,7 +24,7 @@ from cleanX.image_work import (
     WhiteEdgeCrop,
     Mean,
 )
-from cleanX.image_work.steps import FourierTransf, ProjectionHorizoVert
+from cleanX.image_work.steps import CleanRotate, FourierTransf, ProjectionHorizoVert
 # from cleanX.image_work.steps import BlackEdgeCrop, WhiteEdgeCrop
 
 
@@ -102,6 +102,20 @@ def test_projectionhorizovert():
         p = create_pipeline(steps=(
             Acquire(),
             ProjectionHorizoVert(),
+            Save(td),
+        ))
+        p.process(src)
+        src_files = set(f for f in os.listdir(src_dir) if f.endswith('.jpg'))
+        dst_files = set(os.listdir(td))
+        assert src_files == dst_files
+
+def test_cleanrotate():
+    src_dir = image_directory
+    with TemporaryDirectory() as td:
+        src = DirectorySource(src_dir)
+        p = create_pipeline(steps=(
+            Acquire(),
+            CleanRotate(),
             Save(td),
         ))
         p.process(src)
