### 2022-01-01 f33a2a3
[fix] serialization in CleanRotate
