### 2022-01-01 11e5b01
image noise matrix function
