+++ cleanX/image_work/__init__.py
@@ -56,6 +56,9 @@ from .image_functions import (
     cut_to_size,
     cut_or_pad,
     rotated_with_max_clean_area,
+    noise_sum_median_blur,
+    noise_sum_cv,
+    blind_noise_matrix,
 
     Rotator,
 )
