### 2022-01-02 aebaf5f
additional noise function, delete mistake placed file
