### 2022-01-02 072ebb3
noise in segmented function and test
