+++ cleanX/image_work/__init__.py
@@ -59,6 +59,8 @@ from .image_functions import (
     noise_sum_median_blur,
     noise_sum_cv,
     blind_noise_matrix,
+    segmented_blind_noise_matrix,
+
 
     Rotator,
 )
