+++ cleanX/image_work/image_functions.py
@@ -48,26 +48,6 @@ from filecmp import cmp
 from pathlib import Path
 
 
-# def simpler_crop(image):
-#     """
-#     Crops an image of a black frame
-
-#     :param image: Image
-#     :type image: Image (JPEG)
-
-#     :return: image cropped of black edges
-#     :rtype: numpy.ndarray
-#     """
-#     nonzero = np.nonzero(image)
-#     y_nonzero = nonzero[0]
-#     x_nonzero = nonzero[1]
-#     # , x_nonzero, _ = np.nonzero(image)
-#     return image[
-#         np.min(y_nonzero):np.max(y_nonzero),
-#         np.min(x_nonzero):np.max(x_nonzero)
-#     ]
-
-
 def crop_np(image_array):
     """
     Crops black edges of an image array
@@ -2449,6 +2429,93 @@ def noise_sum_median_blur(image):
     return final_sum
 
 
+def noise_sum_gaussian(image):
+    """
+    Given an image, will try to sum up the noise, then divide by the area of
+    the image. The noise sumation here is based on a gaussian filter denoising
+
+    :param img: original image (3 or single channel)
+    :type img: numpy.ndarray
+
+    :return: final_sum
+    :rtype: float
+    """
+    un_noise = cv2.GaussianBlur(image, (3, 3), 0)
+    difference = abs(un_noise - image)
+    sumed = difference.sum()
+    ht, wt = image.shape[0:2]
+    area = ht*wt
+    final_sum = sumed/area
+    return final_sum
+
+
+def noise_sum_bilateral(image):
+    """
+    Given an image, will try to sum up the noise, then divide by the area of
+    the image. The noise sumation here is based on a bilatera filter denoising
+    given a fairly large area (15 pixels)
+
+    :param img: original image (3 or single channel)
+    :type img: numpy.ndarray
+
+    :return: final_sum
+    :rtype: float
+    """
+
+    un_noise = cv2.bilateralFilter(image, 15, 75, 75)
+    difference = abs(un_noise - image)
+    sumed = difference.sum()
+    ht, wt = image.shape[0:2]
+    area = ht*wt
+    final_sum = sumed/area
+    return final_sum
+
+
+def noise_sum_bilateralLO(image):
+    """
+    Given an image, will try to sum up the noise, then divide by the area of
+    the image. The noise sumation here is based on a bilatera filter denoising
+    given a fairly large area (15 pixels)
+
+    :param img: original image (3 or single channel)
+    :type img: numpy.ndarray
+
+    :return: final_sum
+    :rtype: float
+    """
+
+    un_noise = cv2.bilateralFilter(image, 5, 5, 5)
+    difference = abs(un_noise - image)
+    sumed = difference.sum()
+    ht, wt = image.shape[0:2]
+    area = ht*wt
+    final_sum = sumed/area
+    return final_sum
+
+
+def noise_sum_5k(image):
+    """
+    Given an image, will try to sum up the noise, then divide by the area of
+    the image. The noise sumation here is based on a median filter denoising
+    using a 5*5 kernel. This kernel is reccomended for picking up moire
+    patterns and other repetitive noise that may be missed by a smaller kernel.
+
+    :param img: original image (3 or single channel)
+    :type img: numpy.ndarray
+
+    :return: final_sum
+    :rtype: float
+    """
+    kernel_size = 5
+    un_noise = cv2.medianBlur(image, kernel_size)
+    difference = abs(un_noise - image)
+    sumed = difference.sum()
+    ht, wt = image.shape[0:2]
+    area = ht*wt
+    final_sum = sumed/area
+    return final_sum
+
+
 def noise_sum_7k(image):
     """
     Given an image, will try to sum up the noise, then divide by the area of
[…]
