### 2022-01-03 6385fd0
improve segmented noise
