### 2022-01-03 6515abe
segmented function correction
