### 2022-01-03 ea4b4b8
[wip] Redesign steps to to process sources themselves
