+++ cleanX/image_work/pipeline.py
@@ -13,96 +13,13 @@ from tempfile import TemporaryDirectory
 from glob import glob
 from queue import Empty
 
-from .steps import Aggregate, Step, Serial, Parallel
+from .steps import Step
 
 import numpy as np
 
 
 class Loop(ValueError):
     pass
-
-
-class MultiSource:
-    """
-    A class to append multiple sources such as :class:`~.GlobSource` or
-    :class:`~.DirectorySource`.  All source classes implement iterator
-    interface.
-    """
-
-    def __init__(self, sources):
-        """
-        Initializes this iterator with multiple sources in a way similar
-        to :code:`itertools.chain()`
-
-        :param sources: A sequence of sources, such as :class:`~.GlobSource`
-                        or :class:`~.DirectorySource`.
-        :type sources: :code:`Sequence`
-        """
-        self.sources = tuple(sources)
-
-    def __iter__(self):
-        """
-        Iterator implementation.
-        """
-        for src in self.sources:
-            for s in src:
-                yield s
-
-
-class GlobSource:
-    """
-    A class that creates an iterator to list all files matching
-    glob pattern.
-    """
-
-    def __init__(self, expression, recursive=False):
-        """
-        Initializes this iterator with the arguments to be passed to
-        :code:`glob.glob()`.
-
-        :param expression: Expression to be passed to :code:`glob.glob()`
-        :type expression: :code:`Union[str, bytes]`
-        :param recursive: Controls the interpretation of :code:`**`
-                          pattern.  If :code:`True`, will interpret it
-                          to mean any number of path fragments.
-        """
-        self.expression = expression
-        self.recursive = recursive
-
-    def __iter__(self):
-        """
-        Iterator implementation.
-        """
-        for s in glob(self.expression, recursive=self.recursive):
-            yield s
-
-
-class DirectorySource:
-    """
-    A class that creates an iterator to look at files in the given
-    directory.
-    """
-
-    def __init__(self, directory, extension='jpg'):
-        """
-        Initializes this iterator.
-
-        :param directory: The directory in which to look for images.
-        :type directory: Must be valid for :code:`os.path.join()`
-        :param extension: A glob pattern for fle extension.  Whether
-                          it is case-sensitive depends on the
-                          filesystem being used.
-        """
-        self.directory = directory
-        self.extension = extension
-
-    def __iter__(self):
-        """
-        Iterator implementation.
-        """
-        exp = os.path.join(self.directory, '*.' + self.extension)
-        for f in glob(exp):
-            yield f
 
 
 class PipelineError(RuntimeError):
@@ -189,7 +106,7 @@ class PSG:
         # mark dependency as resolved and return the unresolved ones
         for v in self.out(dep):
             self._ready[v] = dep
-        self._in_flight.pop(dep)
+        self._in_flight.remove(dep)
         for d in deps:
             if d != dep:
                 yield d
@@ -206,13 +123,20 @@ class PSG:
 
     def pending(self, dep):
         # True, if dep is either in-flight, or unresolved
-        for v in self._ready.values():
-            if v == dep:
-                return False
-        return True
+        return dep not in set(self._ready.values())
 
     def inflight(self):
         yield from self._in_flight
+
+    def used(self, var):
+        for s in self._def.steps.values():
+            if var not in s.variables:
+                continue
+            if self.pending(s):
+                return True
+            if s in self._in_flight:
+                return True
+        return False
 
 
 class Worker(Process):
@@ -251,28 +175,24 @@ class Worker(Process):
         )
 
     def run(self):
-        print('Starting:', os.getpid())
         while True:
             try:
                 out, rec, td = self.inbound.get_nowait()
                 if rec is None:
-                    print('Terminating (1):', os.getpid())
                     break
[…]
