### 2022-01-04 1234b8b
minor function to pick up inverts and test
