### 2022-01-05 drcandacemakedamoore/cleanX#35 @drcandacemakedamoore
Hi, @henrykironde we have not heard from you or anyone else at JOSS in a while. We have kept the main branch unchanged (except a few lines to deal with some security issues ) but now feel we made some substantial process on the develop branch. Do you think the review should take much longer? We can freeze the develop branch as well if this helps.
