### 2022-01-05 drcandacemakedamoore/cleanX#35 @henrykironde
Thanks for the update @drcandacemakedamoore I think we were waiting for a confirmation from you, that everything is ready to review again. Reviews usually take a short time (2-3 weeks). In Cleanx's case, we could not install the software to test it. Since you have reminded us, I will go ahead and test it.

> We have kept the main branch unchanged....but now feel we made some substantial process on the develop branch.

Authors should update their branches based on the Joss requests(issues), hence no need to freeze the branch
