### 2022-01-07 drcandacemakedamoore/cleanX#48 @henrykironde
Language cleanup and typos
CleanX uses some sensitive language that may offend some users. I would recommend that you remove words like `idiots` since it is against the code of conduct for Joss.

There are typos in the doc strings, like """This class allows normalization by throwing off exxtreme values on" It would be nice to look through the doc strings and try to remove the typos.

Note: I am still failing to install CleanX, but I think it is some complications with my Conda setup. I will keep you updated. My target is to finish with the review and final decision in 14 days.

Ref: openjournals/joss-reviews#3632
