### 2022-01-07 drcandacemakedamoore/cleanX#48 @drcandacemakedamoore
Thank you! I will run over all the doc-strings for typos before midnight
today, and try to remove any offensive language. O.S., please see if there
is anything we can do to help with installation.

On Fri, Jan 7, 2022 at 8:34 AM henry senyondo ***@***.***>
wrote:

> CleanX uses some sensitive language that may offend some users. I would
> recommend that you remove words like idiots since it is against the code
> of conduct for Joss.
>
> There are typos in the doc strings, like """This class allows
> normalization by throwing off exxtreme values on" It would be nice to look
> through the doc strings and try to remove the typos.
>
> Note: I am still failing to install CleanX, but I think it is some
> complications with my Conda setup. I will keep you updated. My target is to
> finish with the review and final decision in 14 days.
>
> Ref: openjournals/joss-reviews#3632
> <https://github.com/openjournals/joss-reviews/issues/3632>
>
