### 2022-01-07 11bf987
fix typos in additional_demos and CLI/dicom.py
