### 2022-01-07 89ca280
typo fixes in dataset_processing and pydicom_adapter
