### 2022-01-07 e1ed61b
fix typos in dicom processing
