### 2022-01-07 722269a
fix image_work init and image_functions to line 600 typos
