### 2022-01-07 5b610e9
remove word idiots
