### 2022-01-07 f374b80
fix typos and reword some docu for image_functions
