+++ cleanX/image_work/image_functions.py
@@ -641,9 +641,9 @@ def find_by_sample_upper(
     """
     This function takes an average (mean) of upper pixels,
     and can show outliers defined by a percentage, i.e. the function shows
-    images with an average of upper pixels in top x % where x is the percent 
+    images with an average of upper pixels in top x % where x is the percent
     height of the sample. Note: images with high averages in the upper pixels
-    are likely to be inverted, upside down or otherwise different from more 
+    are likely to be inverted, upside down or otherwise different from more
     typical X-rays.
 
     :param source_directory: folder the images are in(should include final '/')
@@ -685,7 +685,7 @@ def find_sample_upper_greater_than_lower(
     percent_height_of_sample
 ):
     """
-    Takes average of upper pixels, average of lower (you define what
+    Takes average of upper pixels, average of lower pixels (you define what
     percent of picture should be considered upper and lower) and compares.
     In a CXR if lower average is greater than upper it may be upside down
     or otherwise bizarre, as the neck is smaller than the abdomen.
@@ -723,8 +723,8 @@ def find_sample_upper_greater_than_lower(
 def find_outliers_by_total_mean(source_directory, percentage_to_say_outliers):
     """
     Takes the average of all pixels in an image, returns a :code:`DataFrame`
-    with those images that are outliers by mean. Should catch some inverted
-    or problem images
+    with those images that are outliers by mean. This function can catch some
+    inverted or otherwise problematic images
 
     :param source_directory: folder images in (include final /)
     :type source_directory: :func:`os.path.join()`
@@ -754,9 +754,12 @@ def find_outliers_by_total_mean(source_directory, percentage_to_say_outliers):
 
 def find_outliers_by_mean_to_df(source_directory, percentage_to_say_outliers):
     """
-    Important note: approximate, and it can by chance cut the group
-    so images with
-    the same mean are in and out of normal range if the knife so falls
+    Takes the average of all pixels in an image, returns a :code:`DataFrame`
+    with those images classified. This function can catch some
+    inverted or otherwise problematic images
+    Important note: approximate, and the function can by chance cut the groups
+    so images with the same mean are in and out of normal range,
+    if the knife so falls
 
     :param source_directory: The folder in which the images are
     :type source_directory: :func:`os.path.join()`
@@ -769,7 +772,6 @@ def find_outliers_by_mean_to_df(source_directory, percentage_to_say_outliers):
     suspects1 = glob.glob(os.path.join(source_directory, '*.[Jj][Pp][Gg]'))
     suspects2 = glob.glob(os.path.join(source_directory, '*.[Jj][Pp][Ee][Gg]'))
     suspects = suspects1 + suspects2
-    # suspects = glob.glob(os.path.join(source_directory, '*.jpg'))
     images, means = [], []
     for pic in suspects:
         example = cv2.imread(pic, cv2.IMREAD_GRAYSCALE)
@@ -780,13 +782,9 @@ def find_outliers_by_mean_to_df(source_directory, percentage_to_say_outliers):
     df.sort_values(by='means', inplace=True)
     df.reset_index(inplace=True, drop=True)
     percentile = int((len(df) / 100) * percentage_to_say_outliers)
-    # lows = df.head(percentile)
-    # highs = df.tail(percentile)
     df['results'] = 'within range'
     df.loc[:percentile, 'results'] = 'low'
     df.loc[len(df) - percentile:, 'results'] = 'high'
-    # whole = [lows, highs]
-    # new_df = pd.concat(whole)
     return(df)
 
 
@@ -932,7 +930,8 @@ def find_suspect_text_by_length(directory, length):
     # this function finds all texts above a specified length
     # (length is number of characters)
     """
-    Finds images with text over a specific length you ask for.
+    Finds images with text over a specific length (of letters, digits,
[…]
