### 2022-01-07 5d74d15
finish fixing typos in image_work
