### 2022-01-07 e26fab1
finish fixing typos in image_work
