### 2022-01-07 9669d12
comment out test install trick
