+++ test/test_cleanx.py
@@ -15,9 +15,9 @@ import pandas as pd
 import matplotlib as plt
 import numpy as np
 
-# remove this before commit
-import sys
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+# # remove this before commit
+# import sys
+# sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
 from cleanX import (
     dataset_processing as csvp,
@@ -199,7 +199,6 @@ def test_histogram_difference_for_inverts():
 def test_histogram_difference_for_inverts_todf():
     histos = iwork.histogram_difference_for_inverts_todf(image_directory)
     assert len(histos) >0
-
 
 def test_dataframe_up_my_pics():
     dfy = iwork.dataframe_up_my_pics(image_directory, 'diagnosis_string')
