+++ cleanX/image_work/steps.py
@@ -9,6 +9,7 @@ import pandas as pd
 import math
 from multiprocessing import Queue, Value
 from queue import Empty
+from glob import glob
 
 import numpy as np
 import cv2
@@ -49,23 +50,24 @@ class Applier:
 
 class Splitter(Applier):
 
-    def __call__(self, output, results, acc):
+    def __call__(self, output, n, results, acc):
         return acc
 
 
 class Tee(Splitter):
 
-    def __call__(self, output, results, acc):
-        if result:
+    def __call__(self, output, n, results, acc):
+        # TODO(wvxvw): This can be done in parallel.
+        if results:
             for o in output:
-                for r in results:
-                    self.step.write(o, r)
+                for r, im in results.items():
+                    self.step.write(os.path.join(o, r), im)
         return acc
 
 
 class RoundRobin(Splitter):
 
-    def __call__(self, output, results, acc):
+    def __call__(self, output, results, n, acc):
         if getattr(self, 'last', None) is None:
             self.last = output[0]
         idx = output.index(self.last)
@@ -75,7 +77,7 @@ class RoundRobin(Splitter):
 
 class Joiner(Applier):
 
-    def __call__(self, inputs, acc):
+    def __call__(self, inputs, n, acc):
         return inputs, acc
 
 
@@ -89,10 +91,11 @@ class Avg(Joiner):
         self.width = width
         self.height = height
 
-    def __call__(self, inputs, acc):
-        # TODO(wvxvw): Deal with the case when len(inputs) == 1.
-        # In this case we don't want any averaging to occur, and
-        # should be albe to do things faster and more precise.
+    def __call__(self, inputs, n, acc):
+        if not inputs:
+            return [], acc
+        if len(inputs) == 1:
+            return inputs, acc
         twidth, theight = 0, 0
         for i in inputs:
             meta = self.step.read_meta(i)
@@ -175,21 +178,48 @@ class Step(metaclass=RegisteredStep):
                         for r, o in zip(res, out):
                             step.write(r, o)
 
-    def apply_joined(self, joined, acc):
-        return joined, acc
+    def ensure_out_dirs(self, out):
+        for o in out:
+            try:
+                os.makedirs(os.path.join(self.workdir, o))
+            except FileExistsError:
+                pass
 
-    def apply_one(self, outputs, i, acc):
-        return self.splitter(
-            outputs,
-            self.apply_joined(*self.joiner(i, acc)),
-            acc,
-        )
+    def apply_image(self, image):
+        return image, None
 
-    def apply(self, outputs, inputs):
+    def apply(self, joined, acc):
+        result = {}
+        for var, name in joined.items():
+            key = os.path.join(var, name)
+            path = os.path.join(self.workdir, key)
+            i, e = self.read(path)
+            if e is None:
+                result[key] = self.apply_image(i)
+            else:
+                result[key] = i, e
+        return result, acc
+
+    def apply_one(self, outputs, i, n, acc):
+        print('apply_one', i)
+        joined, acc = self.joiner(i, n, acc)
+        return self.splitter(outputs, n, *self.apply(joined, acc))
+
+    def apply_split_join(self, outputs, inputs):
+        self.ensure_out_dirs(outputs)
         if self.serial:
             acc = None
-            for n, i in enumerate(inputs):
-                acc = self.apply_one(outputs, i, len(inputs) - n, acc)
+            if inputs:
+                print('inputs:', inputs)
+                variables = tuple(inputs.keys())
+                items = zip(*tuple(inputs.values()))
+                for n, i in enumerate(items):
+                    one = dict(zip(variables, i))
+                    acc = self.apply_one(outputs, one, len(inputs) - n, acc)
+            else:
[…]
