+++ test/test_pipeline.py
@@ -66,18 +66,10 @@ def test_alter_images():
                     splitter=None,
                     joiner=None,
                 ),
-                'acquire': StepCall(
-                    definition=Acquire,
-                    options=(),
-                    variables=('src',),
-                    serial=True,
-                    splitter=None,
-                    joiner=None,
-                ),
                 'blur': StepCall(
                     definition=BlurEdges,
                     options=(),
-                    variables=('acquire',),
+                    variables=('src',),
                     serial=True,
                     splitter=None,
                     joiner=None,
