### 2022-01-09 72c5665
[fix] Account for incomplete table in show_images_in_df
