### 2022-01-09 474d599
[fix] Add a cv2.imread wrapper that raises error when it returns None
