### 2022-01-14 drcandacemakedamoore/cleanX#48 @drcandacemakedamoore
Hi  @henrykironde  , if I understand you correctly we do not need to keep the branches frozen. I am therefore going to push a new version very soon so our main branch can be clean of potentially offensive comments, as well as include the new groups of functions on image quality and noise.
