### 2022-01-16 4a83815
[fix] image and delete bindier directory
