+++ README.md
@@ -1,6 +1,6 @@
 <p align="center">
     <img style="width: 30%; height: 30%"
-        src="https://github.com/drcandacemakedamoore/cleanX/blob/main/test/cleanXpic.png">
+        src="https://github.com/drcandacemakedamoore/cleanX/blob/main/cleanXpic.png">
 </p>
 
 # cleanX
