+++ binder/requirements.txt
@@ -1,8 +1,0 @@
-
-cleanX
-numpy
-matplotlib
-pandas
-opencv-python-headless
-tesserocr
-Pillow
