+++ .github/workflows/on-commit.yml
@@ -57,6 +57,7 @@ jobs:
       - uses: actions/checkout@v2
         with:
           fetch-depth: 0
+      - run: sudo apt-get update -y
       - run: >-
           sudo apt-get install -y
           libleptonica-dev
