### 2022-01-16 9511c2f
[fix] genconda command changed
