### 2022-01-16 c4fa374
[fix] anaconda command disappeared from CI images
