+++ .github/workflows/on-tag.yml
@@ -55,15 +55,22 @@ jobs:
         with:
           conda-channels: conda-forge
           python-version: ${{ matrix.python-version }}
-      - run: conda install conda-build anaconda-client conda-verify
+      - run: >
+          conda install
+          conda-build
+          anaconda-client
+          conda-verify
+          anaconda
+          $(conda --version | tr ' ' '=')
+          python=${{ matrix.python-version }}
       - run: conda create --name cleanx-build python=${{ matrix.python-version }}
       - run: conda run -n cleanx-build python setup.py genconda
       - run: conda build -c conda-forge --output-folder ./dist ./conda-pkg/
-      - run: >
+      - run: >-
           conda run -n cleanx-build
           python setup.py anaconda_upload
-          --token ${{ secrets.ANACONDA_TOKEN }}
-          --package ./dist/*/*.tar.bz2
+          --token=${{ secrets.ANACONDA_TOKEN }}
+          --package=./dist/*/*.tar.bz2
       - uses: actions/upload-artifact@v2
         with:
           name: conda-build-${{ matrix.os }}-${{ matrix.python-version }}
