### 2022-01-18 drcandacemakedamoore/cleanX#48 @drcandacemakedamoore
OK, a new less offensive version is on the main... we are currently reviewing user installation issues @henrykironde, so if you have any additional notes on that please let us know.
