### 2022-01-18 drcandacemakedamoore/cleanX#48 @henrykironde
Thank you for the updates, `pip install cleanX` works fine now on Mac and ubuntu.

Please add instructions on how to install `tesseract` or let the users know that `tesseract` is required before installing cleanX.  On Mac I had to run `brew install tesseract`.
