### 2022-01-19 d0a8149
Use global values
dicomfile_directory1 and resources moved up
