### 2022-01-19 349b983
Move common values to the top
dicomfile_directory1, sample_image
