+++ test/test_cleanx.py
@@ -29,6 +29,11 @@ from harness import skip_if_missing
 
 image_directory = os.path.join(os.path.dirname(__file__), 'directory')
 target_directory = os.path.join(os.path.dirname(__file__), 'target')
+dicomfile_directory1 = os.path.join(
+    os.path.dirname(__file__),
+    'dicom_example_folder',
+)
+sample_image = os.path.join(image_directory, 'testtocrop.jpg')
 
 
 def gen_row(columns, rnd):
@@ -57,37 +62,33 @@ def gen_json(directory, name, rows, columns, rnd):
         json.dump(data, jsonfile)
 
 def test_crop():
-    example = cv2.imread(os.path.join(image_directory, 'testtocrop.jpg'), cv2.IMREAD_GRAYSCALE)
+    example = cv2.imread(sample_image, cv2.IMREAD_GRAYSCALE)
     cropped_example = iwork.crop(example)
     assert cropped_example.shape < example.shape
 
 def test_crop_np_white():
-    example = cv2.imread(os.path.join(image_directory, 'testtocrop.jpg'), cv2.IMREAD_GRAYSCALE)
+    example = cv2.imread(sample_image, cv2.IMREAD_GRAYSCALE)
     cropped_example = iwork.crop_np_white(example)
     # our algorithm cuts off a row and column so... we can't call ==
     assert type(cropped_example) is np.ndarray
 
 def test_blur_out_edges():
-    image = os.path.join(image_directory, 'testtocrop.jpg')
-    defblur = iwork.blur_out_edges(image)
+    defblur = iwork.blur_out_edges(sample_image)
     assert type(defblur) == np.ndarray 
 
 
 def test_subtle_sharpie_enhance():
-    image = os.path.join(image_directory, 'testtocrop.jpg')
-    lo = iwork.subtle_sharpie_enhance(image)
+    lo = iwork.subtle_sharpie_enhance(sample_image)
     assert lo.shape[0] >1
 
 
 def harsh_sharpie_enhance():
-    image = os.path.join(image_directory, 'testtocrop.jpg')
-    ho = iwork.harsh_sharpie_enhance(image)
+    ho = iwork.harsh_sharpie_enhance(sample_image)
     assert ho.shape[0] >1
 
 
 def test_salting():
-    lindo_image = os.path.join(image_directory, 'testtocrop.jpg')
-    salt = iwork.salting(lindo_image)
+    salt = iwork.salting(sample_image)
     assert salt.shape[0] > 1
 
 def test_check_paths_for_group_leakage():
@@ -139,8 +140,7 @@ def test_find_very_hazy():
 
 
 def test_show_major_lines_on_image():
-    pic_name1 = os.path.join(image_directory, 'testtocrop.jpg')
-    deflop = iwork.show_major_lines_on_image(pic_name1)
+    deflop = iwork.show_major_lines_on_image(sample_image)
     assert deflop # needs a much better test
 
 
@@ -206,9 +206,7 @@ def test_dataframe_up_my_pics():
     
 
 def test_simple_spinning_template():
-    vovo = os.path.join(image_directory, 'testtocrop.jpg')
-    picy1 = vovo
-    img = cv2.imread(vovo, cv2.IMREAD_GRAYSCALE)
+    img = cv2.imread(sample_image, cv2.IMREAD_GRAYSCALE)
     greys_template1 = img[90:110, 200:350]
     angle_start1 = 0
     angle_stop1 = 30
@@ -224,9 +222,7 @@ def test_simple_spinning_template():
 
 
 def test_def_make_contour_image():
-    vovo = os.path.join(image_directory, 'testtocrop.jpg')
-    picy1 = vovo
-    defMkcont = iwork.make_contour_image(picy1)
[…]
