### 2022-01-19 d8a68ad
Update test_cleanx.py
