### 2022-01-21 6f725e9
Add more installation instructions to the readme
