+++ base-packages.txt
@@ -1,0 +1,58 @@
+# packages in environment at D:\bin\anaconda3\envs\cleanx:
+#
+# Name                    Version                   Build  Channel
+brotlipy                  0.7.0           py37hcc03f2d_1003    conda-forge
+bzip2                     1.0.8                h8ffe710_4    conda-forge
+ca-certificates           2021.10.8            h5b45459_0    conda-forge
+certifi                   2021.10.8        py37h03978a9_1    conda-forge
+cffi                      1.15.0           py37hd8e9650_0    conda-forge
+charset-normalizer        2.0.9              pyhd8ed1ab_0    conda-forge
+click                     8.0.3            py37h03978a9_1    conda-forge
+colorama                  0.4.4              pyh9f0ad1d_0    conda-forge
+conda                     4.11.0           py37h03978a9_0    conda-forge
+conda-package-handling    1.7.3            py37ha54c9ec_1    conda-forge
+cryptography              36.0.1           py37h65266a2_0    conda-forge
+cycler                    0.11.0             pyhd8ed1ab_0    conda-forge
+icu                       68.2                 h0e60522_0    conda-forge
+idna                      3.1                pyhd3deb0d_0    conda-forge
+importlib-metadata        4.10.0           py37h03978a9_0    conda-forge
+jpeg                      9d                   h8ffe710_0    conda-forge
+kiwisolver                1.3.2            py37h8c56517_1    conda-forge
+lerc                      3.0                  h0e60522_0    conda-forge
+libdeflate                1.8                  h8ffe710_0    conda-forge
+lz4-c                     1.9.3                h8ffe710_1    conda-forge
+lzo                       2.10              he774522_1000    conda-forge
+menuinst                  1.4.18           py37h03978a9_1    conda-forge
+munkres                   1.1.4              pyh9f0ad1d_0    conda-forge
+olefile                   0.46               pyh9f0ad1d_1    conda-forge
+openssl                   1.1.1l               h8ffe710_0    conda-forge
+packaging                 21.3               pyhd8ed1ab_0    conda-forge
+pip                       21.3.1             pyhd8ed1ab_0    conda-forge
+pycosat                   0.6.3           py37hcc03f2d_1009    conda-forge
+pycparser                 2.21               pyhd8ed1ab_0    conda-forge
+pyopenssl                 21.0.0             pyhd8ed1ab_0    conda-forge
+pyparsing                 3.0.6              pyhd8ed1ab_0    conda-forge
+pysocks                   1.7.1            py37h03978a9_4    conda-forge
+python                    3.7.0             hc182675_1006    conda-forge
+python-dateutil           2.8.2              pyhd8ed1ab_0    conda-forge
+python_abi                3.7                     2_cp37m    conda-forge
+pywin32                   303              py37hcc03f2d_0    conda-forge
+requests                  2.27.0             pyhd8ed1ab_0    conda-forge
+ruamel_yaml               0.15.80         py37hcc03f2d_1006    conda-forge
+setuptools                60.2.0           py37h03978a9_0    conda-forge
+six                       1.16.0             pyh6c4a22f_0    conda-forge
+sqlite                    3.37.0               h8ffe710_0    conda-forge
+tk                        8.6.11               h8ffe710_1    conda-forge
+tqdm                      4.62.3             pyhd8ed1ab_0    conda-forge
+typing_extensions         4.0.1              pyha770c72_0    conda-forge
+ucrt                      10.0.20348.0         h57928b3_0    conda-forge
+unicodedata2              14.0.0           py37hcc03f2d_0    conda-forge
+urllib3                   1.26.7             pyhd8ed1ab_0    conda-forge
[…]
