### 2022-01-24 e9bc5bc
Update medical-professionals.rst
