+++ docs/medical-professionals.rst
@@ -2,7 +2,7 @@
 Medical Professional Documentation
 ==================================
 
-This is the present home of documentation **for medical
+This is the present documentation home **for medical
 professionals** who do not spend most of their time coding.  From here
 you can link to our `Jupyter notebook`_ specifically for medical
 professionals (one of a series of notebooks), `a brief introductory
@@ -15,37 +15,35 @@ Introduction
 `cleanX` is a package meant to ease the preparation of data for
 machine learning or other algorithms created with some types of
 radiology imaging data.  The package was developed specifically for
-chest X-rays, however, some of the functionality can be used with
+chest X-rays; however, some of Cleanx's functionality can be used with
 other types of data.  For example, the data_processing module can be
-used with many types of tabular data for radiology or even other
-areas.
+used with tabular data for radiology or other areas.
 
-Programmers have a saying "garbage in, garbage out".  If we apply this
-saying here it means if you feed a machine learning algorithm
-incorrect and/or mislabeled data, you should expect incorrect outputs
-(the algorithm will not work as desired).  Therefore the work of
-radiologists and allied professionals is critical in creating machine
-learning algorithms for the field.
+Programmers have a saying, "garbage in, garbage out". That is to say,
+if you feed a machine learning algorithm incorrect and/or mislabeled data, 
+you should expect incorrect outputs (the algorithm will not work as desired).  
+Therefore the work of radiologists and allied professionals such as radiographers
+is critical  in creating machine learning algorithms for the field.
 
-At the present state of the art datasets often include over 100,000
+At present, state of the art datasets often include over 100,000
 images.  The task of reading this many images can be aided by using a
-tool such as `cleanX`.  It will not only automatically reveal some
-questionable images, but also reveal images that are problematic for
+tool such as `cleanX`.  It will automatically reveal some
+questionable images, whilst revealing images that are problematic for
 computer vision programs e.g. inverted or upside-down images that a
 human reader can read easily.
 
-`cleanX` can also be used by anyone (even programmers) to get rid of
-images which obviously should not be in a dataset.  There are public
+`cleanX` can also be used by anyone (even programmers) to eliminate images
+that obviously should not be in a dataset. There are public
 datasets of chest X-rays that contain occasional coronal CT slices, or
 other types of images.  Such images should ideally be thrown out of a
-dataset before you begin reading.  `cleanX` can help anyone with good
+dataset before reading commences.  `cleanX` can help anyone with good
 vision accomplish this task.
 
-`cleanX` is also made so it can extract and clean the metadata on the
-images.  But if you are an imaging professional not very thrilled
-about coding, but still want to make a substantial contribution to a
-machine learning project, then data curation is one of the most
-important tasks you can help with.  We have a video, `our video for
+`cleanX` is also desinged to extract and clean the metadata on 
+images. Imaging professionals that are not proficient
+in coding, but still want to make a substantial contribution to a
+machine learning project, can add value via data curation, one of the most
+important tasks in the project.  We have a video, `our video for
 non-coders demo`_, that shows how `cleanX` can help with some of these
 tasks.
 
@@ -59,7 +57,7 @@ Workflow
 - Image work
 
 To see some functionality of `cleanX`, it is suggested to run **both**
-notebooks in the `workflow demo`_ folder.  If you do not have or know
+notebooks in the `workflow demo`_ folder.  If you do not have, or know
 how to comfortably operate Jupyter notebooks, you can check out our
 videos `video demo of several classes and functions`_ and `video for
 non-coders demo`_
@@ -71,7 +69,7 @@ What makes `cleanX` different and/or important?
 
 Much research about biomedical imaging relies on a combination of
[…]
