### 2022-01-24 37269a2
fix zero_to_twofivefive_simplest_norming function
