+++ cleanX/image_work/image_functions.py
@@ -1564,7 +1564,8 @@ def zero_to_twofivefive_simplest_norming(img_pys):
     min_value = np.amin(img_py)
 
     img_py = img_py - min_value
-    multiplier_ratio = new_max_value/max_value
+    # multiplier_ratio = new_max_value/max_value (mistake from older version)
+    multiplier_ratio = new_max_value/(max_value - min_value)
     img_py = img_py*multiplier_ratio
 
     return img_py
