### 2022-01-24 c41e509
fix steps normalization bug
