### 2022-01-24 82f3b72
[fix] rescale_range_from_histogram_low_end
