### 2022-01-24 32c73af
add explanation on fixed normalize function
