### 2022-01-24 15e5f41
Port changes from image_functions to steps, fix failing test
