### 2022-01-24 8c3edbe
add explanation on fixed normalize function
