### 2022-01-25 43abb49
Try different solver
