### 2022-01-25 779aef0
Some info for bug report
