+++ .github/workflows/on-commit.yml
@@ -96,6 +96,8 @@ jobs:
         with:
           conda-channels: conda-forge
           python-version: ${{ matrix.python-version }}
+      - run: conda config --show-sources
+      - run: conda list --show-channel-urls
       - run: >-
           conda install
           --override-channels
