### 2022-01-25 a3f65ca
Remove the alternative solver (doesn't work anyways)
