+++ .github/workflows/on-commit.yml
@@ -98,6 +98,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - run: conda config --show-sources
       - run: conda list --show-channel-urls
+      - run: conda config --remove channels defaults
       - run: >-
           conda create
           --override-channels
