### 2022-01-29 cf07e29
Replace PNG logo with SVG one
