### 2022-02-20 197740e
linted new helper function for inverting
