### 2022-02-20 dcb3e82
step for inverting all images
