+++ cleanX/image_work/__init__.py
@@ -95,6 +95,7 @@ from .steps import (
     OtsuLines,
     Projection,
     CleanRotate,
+    InvertImages,
 )
 
 
