### 2022-02-21 03685bd
add documentation for make_inverted
