### 2022-02-22 b6c7233
Set strict version requirements for dependencies in Lunux Py3.8
