+++ .github/workflows/on-commit.yml
@@ -107,9 +107,17 @@ jobs:
           -c conda-forge
           python=${{ matrix.python-version }}
       - run: conda run -n cleanx-build python setup.py bdist_egg
-      - run: conda run -n cleanx-build python setup.py genconda
+      - name: Generate conda meta.yaml (Python 3.8)
+        if: ${{ matrix.os == 'ubuntu-latest' && matrix.python-version == '3.8' }}
+        run: conda run -n cleanx-build python setup.py genconda
+        env:
+          STRICT_PACKAGE_VERSIONS: 1
+      - name: Generate conda meta.yaml
+        if: ${{ matrix.os != 'ubuntu-latest' || matrix.python-version != '3.8' }}
+        run: conda run -n cleanx-build python setup.py genconda
         # conda-build installation and the build cannot be in an
         # environment
+      - run: cat ./conda-pkg/meta.yaml
       - run: >-
           conda install
           -yq
