### 2022-02-26 61e7d94
new compare function with hash
