### 2022-03-06 c9bc95c
[fix] Let's see if we can publish w/o anaconda package
