### 2022-03-06 ca5b146
[fix] Copy-paste error in histogram normalize
