### 2022-03-17 #3632 @wvxvw
Hi, Henry, sorry it's taking longer than I hoped.

How about 13:00 UTC on Saturday? If that's inconvenient, then maybe 20:00
UTC on the same day, or it could be an hour or two later.

On Mon, Mar 14, 2022 at 12:09 AM henry senyondo ***@***.***>
wrote:

> ES, any time works for me, night or day.
>
