### 2022-03-18 #3632 @henrykironde
20:00 UTC works fine
