### 2022-03-25 daee854
Create Docker image and push it to DockerHub
