### 2022-03-25 7cead95
Create Docker image and push it to DockerHub
