### 2022-03-27 03f92c4
[doc] Information about running in Docker container
