### 2022-05-08 f02024b
Add Docker image for Jupyter examples
