+++ README.md
@@ -176,6 +176,41 @@ require root privileged.  In the future, it's possible that the image
 will come with a non-root user and will default to running as a
 non-root user.
 
+Additionally, there is a Docker image with several examples in a form
+of Jupyter notebooks.  To run this image:
+
+``` sh
+docker run --rm -ti -p 8888:8888 --network=host \
+    drcandacemakedamoore/cleanx-jupyter-examples
+```
+
+This will generate output similar to:
+
+``` sh
+[I 12:59:52.383 NotebookApp] Writing notebook server cookie secret  \
+to /home/jupyter/.local/share/jupyter/runtime/notebook_cookie_secret
+[I 12:59:52.704 NotebookApp] Serving notebooks from local directory:\
+/home/jupyter
+[I 12:59:52.704 NotebookApp] Jupyter Notebook 6.4.11 is running at:
+[I 12:59:52.705 NotebookApp] http://localhost:8888/?token=...
+[I 12:59:52.705 NotebookApp]  or http://127.0.0.1:8888/?token=...
+[I 12:59:52.705 NotebookApp] Use Control-C to stop this server and \
+shut down all kernels (twice to skip confirmation).
+[W 12:59:52.709 NotebookApp] No web browser found: could not locate\
+runnable browser.
+[C 12:59:52.709 NotebookApp] 
+
+    To access the notebook, open this file in a browser:
+        file:///.../nbserver-1-open.html
+    Or copy and paste one of these URLs:
+        http://localhost:8888/?token=...
+     or http://127.0.0.1:8888/?token=...
+```
+
+Copy the text that starts with `http://127.0.0.1:8888` (including the
+token) and paste it into your browser's address bar.  The demos should
+be fully operational (you may interact with them, re-evaluate them,
+change available parameters etc.)
 
 ### CLI Example
 
