### 2022-05-08 1df8565
Add Docker image for Jupyter examples
