+++ .github/workflows/on-commit.yml
@@ -215,3 +215,37 @@ jobs:
           file: ./docker/Dockerfile
           push: true
           tags: ${{ secrets.DOCKERHUB_USERNAME }}/cleanx:latest
+      - name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          file: ./docker/JupyterDockerfile
+          push: true
+          tags: ${{ secrets.DOCKERHUB_USERNAME }}/cleanx-jupyter-examples:latest
+
+  DockerDev:
+    runs-on: ubuntu-latest
+    if: |
+      github.ref == 'refs/heads/develop' ||
+      github.ref == 'refs/heads/wvxvw/docker'
+    steps:
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to DockerHub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          file: ./docker/Dockerfile
+          push: true
+          tags: ${{ secrets.DOCKERHUB_USERNAME }}/cleanx:develop
+      - name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          file: ./docker/JupyterDockerfile
+          push: true
+          tags: ${{ secrets.DOCKERHUB_USERNAME }}/cleanx-jupyter-examples:develop
