### 2022-05-14 145423c
kill hardcoded paths
