### 2022-05-19 #3632 @henrykironde
@drcandacemakedamoore Thank you for working on the suggestions and issues that were raised. I have tested the software, specifically using the Linux Docker environment since I had issues with my local MacOS environment. At this point, I recommend accepting the software. I am handing it over to @anki-xyz and @cMadan for further review.
