### 2022-06-15 #3632 @anki-xyz
I was going through the software, seems to work on my end. I would also be happy to accept the software as is.
The manuscript could be improved in two (minor) ways:
- Specify **how** CleanX speeds up software
- The manuscript contains abundant "many", and "some" - if you could be more specific here.
