### 2022-06-15 c0cfbad
further paper clean-up
