### 2022-06-15 a47a101
further paper tweaks
