+++ paper/paper.bib
@@ -42,6 +42,21 @@ numpages = {7}
 }
 
 
+@article{DIAZ202125,
+title = {Data preparation for artificial intelligence in medical imaging: A comprehensive guide to open-access platforms and tools},
+journal = {Physica Medica},
+volume = {83},
+pages = {25-37},
+year = {2021},
+issn = {1120-1797},
+doi = {https://doi.org/10.1016/j.ejmp.2021.02.007},
+url = {https://www.sciencedirect.com/science/article/pii/S1120179721000958},
+author = {Oliver Diaz and Kaisar Kushibar and Richard Osuala and Akis Linardos and Lidia Garrucho and Laura Igual and Petia Radeva and Fred Prior and Polyxeni Gkontra and Karim Lekadir},
+keywords = {Artificial intelligence, Medical image datasets, Data curation, Annotation, Segmentation, Image storage, Open access, Image quality assessment},
+abstract = {The vast amount of data produced by today’s medical imaging systems has led medical professionals to turn to novel technologies in order to efficiently handle their data and exploit the rich information present in them. In this context, artificial intelligence (AI) is emerging as one of the most prominent solutions, promising to revolutionise every day clinical practice and medical research. The pillar supporting the development of reliable and robust AI algorithms is the appropriate preparation of the medical images to be used by the AI-driven solutions. Here, we provide a comprehensive guide for the necessary steps to prepare medical images prior to developing or applying AI algorithms. The main steps involved in a typical medical image preparation pipeline include: (i) image acquisition at clinical sites, (ii) image de-identification to remove personal information and protect patient privacy, (iii) data curation to control for image and associated information quality, (iv) image storage, and (v) image annotation. There exists a plethora of open access tools to perform each of the aforementioned tasks and are hereby reviewed. Furthermore, we detail medical image repositories covering different organs and diseases. Such repositories are constantly increasing and enriched with the advent of big data. Lastly, we offer directions for future work in this rapidly evolving field.}
+}
+
+
 @Inbook{Harvey2019,
 author={Harvey H., Glocker B.},
 editor={Ranschaert E., Morozov S., Algra P. (eds)}, 
