### 2022-06-15 8868ee8
Cosmetic changes to paper (few commas and misplaced hyphens) + formatting
