+++ paper/paper.md
@@ -38,12 +38,12 @@ bibliography: paper.bib
 
 
 Radiological images of various anatomy are part of the diagnostic
-work-up for millions of patients for diverse indications. A
-considerable amount of time and resources have gone into 
-developing automated diagnostic interpretation of these images. The
-purpose of this library is to help scientists, medical professionals,
-and programmers create better datasets upon which algorithms related
-to X-rays, MRIs or CTs can be based.
+work-up for millions of patients for diverse indications.  A
+considerable amount of time and resources have gone into developing
+automated diagnostic interpretation of these images.  The purpose of
+this library is to help scientists, medical professionals, and
+programmers create better datasets upon which algorithms related to
+X-rays, MRIs or CTs can be based.
 
 CleanX is a Python package for data cleaning that was developed for
 radiology artificial intelligence (AI).
@@ -51,71 +51,81 @@ radiology artificial intelligence (AI).
 
 # Statement of need
 `CleanX` is a Python package for data exploration, cleaning, and
-augmentation that was originally developed for radiology AI. Python is
-a widely used language on a global level. Data preparation for
-building quality machine learning algorithms is a
-time-consuming task [@10.1145/2207243.2207253]. Of the tasks involved,
-'data cleaning' alone usually takes the majority of time spent on
-analysis for clinical research projects [@tidy-data]. Even in the case of relatively high-quality datasets, the task of 'cleaning' is a necessary step, to avoid the problem of poor input leading to poor performance, also known as the "garbage in, garbage out" phenomenon [@Rahm2000DataCP].
+augmentation that was originally developed for radiology AI.  Python
+is a widely used language on a global level.  Data preparation for
+building quality machine learning algorithms is a time-consuming task
+[@10.1145/2207243.2207253].  Of the tasks involved, 'data cleaning'
+alone usually takes the majority of time spent on analysis for
+clinical research projects [@tidy-data].  Even in the case of
+relatively high-quality datasets, the task of 'cleaning' is a
+necessary step, to avoid the problem of poor input leading to poor
+performance, also known as the "garbage in, garbage out" phenomenon
+[@Rahm2000DataCP].
 
 In contemporary research, many approaches to data cleaning for
-radiology datasets overlook the content of the images themselves. Recent work
-reveals more open algorithms for image de-identification than for all
-kinds of image curation[@DIAZ202125]. However the quality of data, especially the image data,
-is often context-specific and salient to a particular AI model.
+radiology datasets overlook the content of the images themselves.
+Recent work reveals more open algorithms for image de-identification
+than for all kinds of image curation[@DIAZ202125].  However the
+quality of data, especially the image data, is often context-specific
+and salient to a particular AI model.
 
 Algorithms that rely on shape detection may be accomplished with
-contrast and positional invariance, but specific neural networks or
+contrast, and positional invariance, but specific neural networks or
 radiomics algorithms should not be insensitive to contrast or
-position. A neural network designed to detect the technical quality of a  
-chest X-ray quality should not be positionally invariant, as a
-the rotated patient is likely to indicate poor radiographic technique,
-and in the case of a flipped image, would correlate with the clinical
-picture of situs inversus. Thus scales like MIDaR [@Harvey2019] are
-necessary but not sufficient to describe data. Despite the specific
+position.  A neural network designed to detect the technical quality
+of a chest X-ray should not be positionally invariant, as a the
+rotated patient is likely to indicate poor radiographic technique, and
+in the case of a flipped image, would correlate with the clinical
+picture of situs inversus.  Thus scales like MIDaR [@Harvey2019] are
+necessary but not sufficient to describe data.  Despite the specific
 nature of quality issues for each model, potential data contamination
-problems should be cleaned out of imaging datasets for most algorithms.
[…]
