+++ paper/paper.bib
@@ -49,7 +49,7 @@ volume = {83},
 pages = {25-37},
 year = {2021},
 issn = {1120-1797},
-doi = {https://doi.org/10.1016/j.ejmp.2021.02.007},
+doi = {10.1016/j.ejmp.2021.02.007},
 url = {https://www.sciencedirect.com/science/article/pii/S1120179721000958},
 author = {Oliver Diaz and Kaisar Kushibar and Richard Osuala and Akis Linardos and Lidia Garrucho and Laura Igual and Petia Radeva and Fred Prior and Polyxeni Gkontra and Karim Lekadir},
 keywords = {Artificial intelligence, Medical image datasets, Data curation, Annotation, Segmentation, Image storage, Open access, Image quality assessment},
