### 2022-08-01 #3632 @andrew-f-murphy
Hey @cMadan totally understand, I have removed https://doi.org/ from that DOI now
